Create Polynomials

Creates polynomial variables, to be used in regression. Will create polynomials of degree less than or equal to the degree specified, and will mean center variables by default.

Usage

polynomial(x, degree = 2, center = mean(x, na.rm = TRUE))

Arguments

x

variable used to create the polynomials.

degree

the maximum degree polynomial to be returned. Polynomials of degree <= degree will be returned.

center

the value to center the polynomials at.

Value

A matrix containing the linear splines.

Examples

# Reading in a dataset
data(mri)

# Create a polynomial on ldl
polynomial(mri$ldl, degree=3)
#>         Linear(ctr)  Square(ctr)     Cube(ctr)
#>   [1,]    9.1972414 8.458925e+01  7.779877e+02
#>   [2,]  -41.8027586 1.747471e+03 -7.304909e+04
#>   [3,]  -10.8027586 1.166996e+02 -1.260678e+03
#>   [4,]  -64.8027586 4.199398e+03 -2.721325e+05
#>   [5,]   22.1972414 4.927175e+02  1.093697e+04
#>   [6,]   37.1972414 1.383635e+03  5.146740e+04
#>   [7,]  -24.8027586 6.151768e+02 -1.525808e+04
#>   [8,]   -9.8027586 9.609408e+01 -9.419870e+02
#>   [9,]   -1.8027586 3.249939e+00 -5.858855e+00
#>  [10,]  -15.8027586 2.497272e+02 -3.946378e+03
#>  [11,]   10.1972414 1.039837e+02  1.060347e+03
#>  [12,]  -17.8027586 3.169382e+02 -5.642375e+03
#>  [13,]   80.1972414 6.431598e+03  5.157964e+05
#>  [14,]    4.1972414 1.761684e+01  7.394211e+01
#>  [15,]   -4.8027586 2.306649e+01 -1.107828e+02
#>  [16,]  -30.8027586 9.488099e+02 -2.922596e+04
#>  [17,]  -47.8027586 2.285104e+03 -1.092343e+05
#>  [18,]  -15.8027586 2.497272e+02 -3.946378e+03
#>  [19,]   10.1972414 1.039837e+02  1.060347e+03
#>  [20,]   34.1972414 1.169451e+03  3.999201e+04
#>  [21,]    4.1972414 1.761684e+01  7.394211e+01
#>  [22,]  -42.8027586 1.832076e+03 -7.841791e+04
#>  [23,]   28.1972414 7.950844e+02  2.241919e+04
#>  [24,]  -55.8027586 3.113948e+03 -1.737669e+05
#>  [25,]  -50.8027586 2.580920e+03 -1.311179e+05
#>  [26,]  -11.8027586 1.393051e+02 -1.644185e+03
#>  [27,]   13.1972414 1.741672e+02  2.298526e+03
#>  [28,]  -44.8027586 2.007287e+03 -8.993200e+04
#>  [29,]   -4.8027586 2.306649e+01 -1.107828e+02
#>  [30,]  -20.8027586 4.327548e+02 -9.002493e+03
#>  [31,]   -3.8027586 1.446097e+01 -5.499159e+01
#>  [32,]    9.1972414 8.458925e+01  7.779877e+02
#>  [33,]   10.1972414 1.039837e+02  1.060347e+03
#>  [34,]   62.1972414 3.868497e+03  2.406098e+05
#>  [35,]   19.1972414 3.685341e+02  7.074838e+03
#>  [36,]  -26.8027586 7.183879e+02 -1.925478e+04
#>  [37,]   45.1972414 2.042791e+03  9.232850e+04
#>  [38,]  -42.8027586 1.832076e+03 -7.841791e+04
#>  [39,]   29.1972414 8.524789e+02  2.489003e+04
#>  [40,]   17.1972414 2.957451e+02  5.086000e+03
#>  [41,]   22.1972414 4.927175e+02  1.093697e+04
#>  [42,]   36.1972414 1.310240e+03  4.742708e+04
#>  [43,]   56.1972414 3.158130e+03  1.774782e+05
#>  [44,]  -43.8027586 1.918682e+03 -8.404355e+04
#>  [45,]   25.1972414 6.349010e+02  1.599775e+04
#>  [46,]  -13.8027586 1.905161e+02 -2.629648e+03
#>  [47,]   33.1972414 1.102057e+03  3.658525e+04
#>  [48,]  -50.8027586 2.580920e+03 -1.311179e+05
#>  [49,]  -27.8027586 7.729934e+02 -2.149135e+04
#>  [50,]   63.1972414 3.993891e+03  2.524029e+05
#>  [51,]   -2.8027586 7.855456e+00 -2.201695e+01
#>  [52,]  101.1972414 1.024088e+04  1.036349e+06
#>  [53,]  -52.8027586 2.788131e+03 -1.472210e+05
#>  [54,]  -26.8027586 7.183879e+02 -1.925478e+04
#>  [55,]   90.1972414 8.135542e+03  7.338035e+05
#>  [56,]  -21.8027586 4.753603e+02 -1.036417e+04
#>  [57,]  -45.8027586 2.097893e+03 -9.608927e+04
#>  [58,]  -36.8027586 1.354443e+03 -4.984724e+04
#>  [59,]  -17.8027586 3.169382e+02 -5.642375e+03
#>  [60,]   21.1972414 4.493230e+02  9.524409e+03
#>  [61,]   17.1972414 2.957451e+02  5.086000e+03
#>  [62,]   25.1972414 6.349010e+02  1.599775e+04
#>  [63,]   25.1972414 6.349010e+02  1.599775e+04
#>  [64,]   12.1972414 1.487727e+02  1.814616e+03
#>  [65,]  -12.8027586 1.639106e+02 -2.098508e+03
#>  [66,]   20.1972414 4.079286e+02  8.239032e+03
#>  [67,]   23.1972414 5.381120e+02  1.248271e+04
#>  [68,]  -53.8027586 2.894737e+03 -1.557448e+05
#>  [69,]    6.1972414 3.840580e+01  2.380100e+02
#>  [70,]  -31.8027586 1.011415e+03 -3.216580e+04
#>  [71,]  -40.8027586 1.664865e+03 -6.793109e+04
#>  [72,]  -46.8027586 2.190498e+03 -1.025214e+05
#>  [73,]    9.1972414 8.458925e+01  7.779877e+02
#>  [74,]   13.1972414 1.741672e+02  2.298526e+03
#>  [75,]  -40.8027586 1.664865e+03 -6.793109e+04
#>  [76,]  -22.8027586 5.199658e+02 -1.185665e+04
#>  [77,]    1.1972414 1.433387e+00  1.716110e+00
#>  [78,]   14.1972414 2.015617e+02  2.861620e+03
#>  [79,]  -24.8027586 6.151768e+02 -1.525808e+04
#>  [80,]   -4.8027586 2.306649e+01 -1.107828e+02
#>  [81,]   20.1972414 4.079286e+02  8.239032e+03
#>  [82,]  -38.8027586 1.505654e+03 -5.842353e+04
#>  [83,]   24.1972414 5.855065e+02  1.416764e+04
#>  [84,]    8.1972414 6.719477e+01  5.508117e+02
#>  [85,]  -45.8027586 2.097893e+03 -9.608927e+04
#>  [86,]    3.1972414 1.022235e+01  3.268333e+01
#>  [87,]   28.1972414 7.950844e+02  2.241919e+04
#>  [88,]  -26.8027586 7.183879e+02 -1.925478e+04
#>  [89,]   -8.8027586 7.748856e+01 -6.821131e+02
#>  [90,]   47.1972414 2.227580e+03  1.051356e+05
#>  [91,]   -5.8027586 3.367201e+01 -1.953905e+02
#>  [92,]   -0.8027586 6.444214e-01 -5.173148e-01
#>  [93,]  -15.8027586 2.497272e+02 -3.946378e+03
#>  [94,]  -40.8027586 1.664865e+03 -6.793109e+04
#>  [95,]  -55.8027586 3.113948e+03 -1.737669e+05
#>  [96,]   41.1972414 1.697213e+03  6.992048e+04
#>  [97,]   -3.8027586 1.446097e+01 -5.499159e+01
#>  [98,]   37.1972414 1.383635e+03  5.146740e+04
#>  [99,]  -25.8027586 6.657824e+02 -1.717902e+04
#> [100,]   32.1972414 1.036662e+03  3.337767e+04
#> [101,]  -10.8027586 1.166996e+02 -1.260678e+03
#> [102,]  -28.8027586 8.295989e+02 -2.389474e+04
#> [103,]  -41.8027586 1.747471e+03 -7.304909e+04
#> [104,]  -44.8027586 2.007287e+03 -8.993200e+04
#> [105,]   25.1972414 6.349010e+02  1.599775e+04
#> [106,]   65.1972414 4.250680e+03  2.771326e+05
#> [107,]   41.1972414 1.697213e+03  6.992048e+04
#> [108,]   -8.8027586 7.748856e+01 -6.821131e+02
#> [109,]  -21.8027586 4.753603e+02 -1.036417e+04
#> [110,]  -12.8027586 1.639106e+02 -2.098508e+03
#> [111,]   -7.8027586 6.088304e+01 -4.750557e+02
#> [112,]   -5.8027586 3.367201e+01 -1.953905e+02
#> [113,]  -27.8027586 7.729934e+02 -2.149135e+04
#> [114,]    0.1972414 3.890416e-02  7.673511e-03
#> [115,]   16.1972414 2.623506e+02  4.249356e+03
#> [116,]   20.1972414 4.079286e+02  8.239032e+03
#> [117,]  -42.8027586 1.832076e+03 -7.841791e+04
#> [118,]    7.1972414 5.180028e+01  3.728191e+02
#> [119,]    2.1972414 4.827870e+00  1.060800e+01
#> [120,]    4.1972414 1.761684e+01  7.394211e+01
#> [121,]  -35.8027586 1.281838e+03 -4.589332e+04
#> [122,]   10.1972414 1.039837e+02  1.060347e+03
#> [123,]  -18.8027586 3.535437e+02 -6.647597e+03
#> [124,]   -2.8027586 7.855456e+00 -2.201695e+01
#> [125,]  -21.8027586 4.753603e+02 -1.036417e+04
#> [126,]  -23.8027586 5.665713e+02 -1.348596e+04
#> [127,]  -25.8027586 6.657824e+02 -1.717902e+04
#> [128,]   51.1972414 2.621158e+03  1.341960e+05
#> [129,]  -39.8027586 1.584260e+03 -6.305790e+04
#> [130,]           NA           NA            NA
#> [131,]   38.1972414 1.459029e+03  5.573089e+04
#> [132,]   30.1972414 9.118734e+02  2.753606e+04
#> [133,]  -17.8027586 3.169382e+02 -5.642375e+03
#> [134,]  -72.8027586 5.300242e+03 -3.858722e+05
#> [135,]   24.1972414 5.855065e+02  1.416764e+04
#> [136,]   69.1972414 4.788258e+03  3.313343e+05
#> [137,]    8.1972414 6.719477e+01  5.508117e+02
#> [138,]   16.1972414 2.623506e+02  4.249356e+03
#> [139,]  -37.8027586 1.429049e+03 -5.402198e+04
#> [140,]    4.1972414 1.761684e+01  7.394211e+01
#> [141,]   48.1972414 2.322974e+03  1.119609e+05
#> [142,]   25.1972414 6.349010e+02  1.599775e+04
#> [143,]   22.1972414 4.927175e+02  1.093697e+04
#> [144,]   -6.8027586 4.627752e+01 -3.148148e+02
#> [145,]   35.1972414 1.238846e+03  4.360395e+04
#> [146,]   10.1972414 1.039837e+02  1.060347e+03
#> [147,]   30.1972414 9.118734e+02  2.753606e+04
#> [148,]  -26.8027586 7.183879e+02 -1.925478e+04
#> [149,]  -11.8027586 1.393051e+02 -1.644185e+03
#> [150,]   12.1972414 1.487727e+02  1.814616e+03
#> [151,]  -21.8027586 4.753603e+02 -1.036417e+04
#> [152,]  -12.8027586 1.639106e+02 -2.098508e+03
#> [153,]    7.1972414 5.180028e+01  3.728191e+02
#> [154,]  -44.8027586 2.007287e+03 -8.993200e+04
#> [155,]  -47.8027586 2.285104e+03 -1.092343e+05
#> [156,]  -10.8027586 1.166996e+02 -1.260678e+03
#> [157,]   15.1972414 2.309561e+02  3.509896e+03
#> [158,]           NA           NA            NA
#> [159,]  -47.8027586 2.285104e+03 -1.092343e+05
#> [160,]  -24.8027586 6.151768e+02 -1.525808e+04
#> [161,]   29.1972414 8.524789e+02  2.489003e+04
#> [162,]   10.1972414 1.039837e+02  1.060347e+03
#> [163,]    5.1972414 2.701132e+01  1.403843e+02
#> [164,]   12.1972414 1.487727e+02  1.814616e+03
#> [165,]   14.1972414 2.015617e+02  2.861620e+03
#> [166,]  -10.8027586 1.166996e+02 -1.260678e+03
#> [167,]   18.1972414 3.311396e+02  6.025827e+03
#> [168,]    0.1972414 3.890416e-02  7.673511e-03
#> [169,]   91.1972414 8.316937e+03  7.584817e+05
#> [170,]   19.1972414 3.685341e+02  7.074838e+03
#> [171,]    1.1972414 1.433387e+00  1.716110e+00
#> [172,]  -49.8027586 2.480315e+03 -1.235265e+05
#> [173,]           NA           NA            NA
#> [174,]  -74.8027586 5.595453e+03 -4.185553e+05
#> [175,]    6.1972414 3.840580e+01  2.380100e+02
#> [176,]  -11.8027586 1.393051e+02 -1.644185e+03
#> [177,]  -64.8027586 4.199398e+03 -2.721325e+05
#> [178,]   23.1972414 5.381120e+02  1.248271e+04
#> [179,]   11.1972414 1.253782e+02  1.403890e+03
#> [180,]    1.1972414 1.433387e+00  1.716110e+00
#> [181,]    7.1972414 5.180028e+01  3.728191e+02
#> [182,]    7.1972414 5.180028e+01  3.728191e+02
#> [183,]   18.1972414 3.311396e+02  6.025827e+03
#> [184,]  -16.8027586 2.823327e+02 -4.743968e+03
#> [185,]   26.1972414 6.862955e+02  1.797905e+04
#> [186,]  -30.8027586 9.488099e+02 -2.922596e+04
#> [187,]    3.1972414 1.022235e+01  3.268333e+01
#> [188,]           NA           NA            NA
#> [189,]    2.1972414 4.827870e+00  1.060800e+01
#> [190,]  -26.8027586 7.183879e+02 -1.925478e+04
#> [191,]   13.1972414 1.741672e+02  2.298526e+03
#> [192,]   16.1972414 2.623506e+02  4.249356e+03
#> [193,]  -22.8027586 5.199658e+02 -1.185665e+04
#> [194,]  -11.8027586 1.393051e+02 -1.644185e+03
#> [195,]   59.1972414 3.504313e+03  2.074457e+05
#> [196,]   -1.8027586 3.249939e+00 -5.858855e+00
#> [197,]  -27.8027586 7.729934e+02 -2.149135e+04
#> [198,]  -24.8027586 6.151768e+02 -1.525808e+04
#> [199,]   65.1972414 4.250680e+03  2.771326e+05
#> [200,]   -7.8027586 6.088304e+01 -4.750557e+02
#> [201,]  -28.8027586 8.295989e+02 -2.389474e+04
#> [202,]   -8.8027586 7.748856e+01 -6.821131e+02
#> [203,]  -30.8027586 9.488099e+02 -2.922596e+04
#> [204,]    9.1972414 8.458925e+01  7.779877e+02
#> [205,]   80.1972414 6.431598e+03  5.157964e+05
#> [206,]   24.1972414 5.855065e+02  1.416764e+04
#> [207,]   -3.8027586 1.446097e+01 -5.499159e+01
#> [208,]   14.1972414 2.015617e+02  2.861620e+03
#> [209,]  -34.8027586 1.211232e+03 -4.215422e+04
#> [210,]   -4.8027586 2.306649e+01 -1.107828e+02
#> [211,]   34.1972414 1.169451e+03  3.999201e+04
#> [212,]   36.1972414 1.310240e+03  4.742708e+04
#> [213,]   62.1972414 3.868497e+03  2.406098e+05
#> [214,]  -30.8027586 9.488099e+02 -2.922596e+04
#> [215,]   38.1972414 1.459029e+03  5.573089e+04
#> [216,]    4.1972414 1.761684e+01  7.394211e+01
#> [217,]   28.1972414 7.950844e+02  2.241919e+04
#> [218,]           NA           NA            NA
#> [219,]   10.1972414 1.039837e+02  1.060347e+03
#> [220,]  -30.8027586 9.488099e+02 -2.922596e+04
#> [221,]  121.1972414 1.468877e+04  1.780239e+06
#> [222,]   18.1972414 3.311396e+02  6.025827e+03
#> [223,]  -42.8027586 1.832076e+03 -7.841791e+04
#> [224,]   -3.8027586 1.446097e+01 -5.499159e+01
#> [225,]   31.1972414 9.732679e+02  3.036327e+04
#> [226,]  -18.8027586 3.535437e+02 -6.647597e+03
#> [227,]    5.1972414 2.701132e+01  1.403843e+02
#> [228,]   14.1972414 2.015617e+02  2.861620e+03
#> [229,]   74.1972414 5.505231e+03  4.084729e+05
#> [230,]  -30.8027586 9.488099e+02 -2.922596e+04
#> [231,]   20.1972414 4.079286e+02  8.239032e+03
#> [232,]  -23.8027586 5.665713e+02 -1.348596e+04
#> [233,]   11.1972414 1.253782e+02  1.403890e+03
#> [234,]   24.1972414 5.855065e+02  1.416764e+04
#> [235,]   22.1972414 4.927175e+02  1.093697e+04
#> [236,]  -53.8027586 2.894737e+03 -1.557448e+05
#> [237,]   16.1972414 2.623506e+02  4.249356e+03
#> [238,]  -34.8027586 1.211232e+03 -4.215422e+04
#> [239,]   27.1972414 7.396899e+02  2.011753e+04
#> [240,]   82.1972414 6.756386e+03  5.553563e+05
#> [241,]   -7.8027586 6.088304e+01 -4.750557e+02
#> [242,]   24.1972414 5.855065e+02  1.416764e+04
#> [243,]   25.1972414 6.349010e+02  1.599775e+04
#> [244,]    1.1972414 1.433387e+00  1.716110e+00
#> [245,]  -25.8027586 6.657824e+02 -1.717902e+04
#> [246,]  -11.8027586 1.393051e+02 -1.644185e+03
#> [247,]  -13.8027586 1.905161e+02 -2.629648e+03
#> [248,]    3.1972414 1.022235e+01  3.268333e+01
#> [249,]   53.1972414 2.829946e+03  1.505453e+05
#> [250,]    8.1972414 6.719477e+01  5.508117e+02
#> [251,]  -10.8027586 1.166996e+02 -1.260678e+03
#> [252,]  -17.8027586 3.169382e+02 -5.642375e+03
#> [253,]  -26.8027586 7.183879e+02 -1.925478e+04
#> [254,]   81.1972414 6.592992e+03  5.353328e+05
#> [255,]  -48.8027586 2.381709e+03 -1.162340e+05
#> [256,]   15.1972414 2.309561e+02  3.509896e+03
#> [257,]    4.1972414 1.761684e+01  7.394211e+01
#> [258,]   36.1972414 1.310240e+03  4.742708e+04
#> [259,]    1.1972414 1.433387e+00  1.716110e+00
#> [260,]  -33.8027586 1.142626e+03 -3.862393e+04
#> [261,]           NA           NA            NA
#> [262,]  -25.8027586 6.657824e+02 -1.717902e+04
#> [263,]   61.1972414 3.745102e+03  2.291899e+05
#> [264,]  -12.8027586 1.639106e+02 -2.098508e+03
#> [265,]    2.1972414 4.827870e+00  1.060800e+01
#> [266,]   57.1972414 3.271524e+03  1.871222e+05
#> [267,]           NA           NA            NA
#> [268,]  -15.8027586 2.497272e+02 -3.946378e+03
#> [269,]  -22.8027586 5.199658e+02 -1.185665e+04
#> [270,]   30.1972414 9.118734e+02  2.753606e+04
#> [271,]  -26.8027586 7.183879e+02 -1.925478e+04
#> [272,]   29.1972414 8.524789e+02  2.489003e+04
#> [273,]   55.1972414 3.046735e+03  1.681714e+05
#> [274,]   -7.8027586 6.088304e+01 -4.750557e+02
#> [275,]    6.1972414 3.840580e+01  2.380100e+02
#> [276,]   22.1972414 4.927175e+02  1.093697e+04
#> [277,]  -35.8027586 1.281838e+03 -4.589332e+04
#> [278,]    5.1972414 2.701132e+01  1.403843e+02
#> [279,]   12.1972414 1.487727e+02  1.814616e+03
#> [280,]  -16.8027586 2.823327e+02 -4.743968e+03
#> [281,]   10.1972414 1.039837e+02  1.060347e+03
#> [282,]   13.1972414 1.741672e+02  2.298526e+03
#> [283,]  -34.8027586 1.211232e+03 -4.215422e+04
#> [284,]   50.1972414 2.519763e+03  1.264852e+05
#> [285,]   22.1972414 4.927175e+02  1.093697e+04
#> [286,]  -38.8027586 1.505654e+03 -5.842353e+04
#> [287,]  -12.8027586 1.639106e+02 -2.098508e+03
#> [288,]   29.1972414 8.524789e+02  2.489003e+04
#> [289,]   59.1972414 3.504313e+03  2.074457e+05
#> [290,]   11.1972414 1.253782e+02  1.403890e+03
#> [291,]   11.1972414 1.253782e+02  1.403890e+03
#> [292,]   43.1972414 1.866002e+03  8.060612e+04
#> [293,]   80.1972414 6.431598e+03  5.157964e+05
#> [294,]  -37.8027586 1.429049e+03 -5.402198e+04
#> [295,]   13.1972414 1.741672e+02  2.298526e+03
#> [296,]   56.1972414 3.158130e+03  1.774782e+05
#> [297,]   11.1972414 1.253782e+02  1.403890e+03
#> [298,]   59.1972414 3.504313e+03  2.074457e+05
#> [299,]   -9.8027586 9.609408e+01 -9.419870e+02
#> [300,]  -11.8027586 1.393051e+02 -1.644185e+03
#> [301,]  -56.8027586 3.226553e+03 -1.832771e+05
#> [302,]   -6.8027586 4.627752e+01 -3.148148e+02
#> [303,]  -43.8027586 1.918682e+03 -8.404355e+04
#> [304,]    4.1972414 1.761684e+01  7.394211e+01
#> [305,]  -15.8027586 2.497272e+02 -3.946378e+03
#> [306,]   36.1972414 1.310240e+03  4.742708e+04
#> [307,]  -12.8027586 1.639106e+02 -2.098508e+03
#> [308,]  -32.8027586 1.076021e+03 -3.529646e+04
#> [309,]   -7.8027586 6.088304e+01 -4.750557e+02
#> [310,]   70.1972414 4.927653e+03  3.459076e+05
#> [311,]  -47.8027586 2.285104e+03 -1.092343e+05
#> [312,]  -14.8027586 2.191217e+02 -3.243605e+03
#> [313,]  -54.8027586 3.003342e+03 -1.645914e+05
#> [314,]   15.1972414 2.309561e+02  3.509896e+03
#> [315,]  -13.8027586 1.905161e+02 -2.629648e+03
#> [316,]   69.1972414 4.788258e+03  3.313343e+05
#> [317,]   -9.8027586 9.609408e+01 -9.419870e+02
#> [318,]   18.1972414 3.311396e+02  6.025827e+03
#> [319,]   47.1972414 2.227580e+03  1.051356e+05
#> [320,]   27.1972414 7.396899e+02  2.011753e+04
#> [321,]   10.1972414 1.039837e+02  1.060347e+03
#> [322,]   56.1972414 3.158130e+03  1.774782e+05
#> [323,]  -35.8027586 1.281838e+03 -4.589332e+04
#> [324,]    2.1972414 4.827870e+00  1.060800e+01
#> [325,]   62.1972414 3.868497e+03  2.406098e+05
#> [326,]   58.1972414 3.386919e+03  1.971093e+05
#> [327,]   -4.8027586 2.306649e+01 -1.107828e+02
#> [328,]   -0.8027586 6.444214e-01 -5.173148e-01
#> [329,]  -12.8027586 1.639106e+02 -2.098508e+03
#> [330,]    8.1972414 6.719477e+01  5.508117e+02
#> [331,]   14.1972414 2.015617e+02  2.861620e+03
#> [332,]  -45.8027586 2.097893e+03 -9.608927e+04
#> [333,]   45.1972414 2.042791e+03  9.232850e+04
#>  [ reached getOption("max.print") -- omitted 402 rows ]
#> attr(,"transformation")
#> [1] "polynomial"
#> attr(,"name")
#> [1] "polynomial(mri$ldl)"
#> attr(,"prnm")
#> [1] "mri$ldl"
#> attr(,"degree")
#> [1] 1 2 3
#> attr(,"center")
#> [1] 125.8028
#> attr(,"original")
#>   [1]    9.1972414  -41.8027586  -10.8027586  -64.8027586   22.1972414
#>   [6]   37.1972414  -24.8027586   -9.8027586   -1.8027586  -15.8027586
#>  [11]   10.1972414  -17.8027586   80.1972414    4.1972414   -4.8027586
#>  [16]  -30.8027586  -47.8027586  -15.8027586   10.1972414   34.1972414
#>  [21]    4.1972414  -42.8027586   28.1972414  -55.8027586  -50.8027586
#>  [26]  -11.8027586   13.1972414  -44.8027586   -4.8027586  -20.8027586
#>  [31]   -3.8027586    9.1972414   10.1972414   62.1972414   19.1972414
#>  [36]  -26.8027586   45.1972414  -42.8027586   29.1972414   17.1972414
#>  [41]   22.1972414   36.1972414   56.1972414  -43.8027586   25.1972414
#>  [46]  -13.8027586   33.1972414  -50.8027586  -27.8027586   63.1972414
#>  [51]   -2.8027586  101.1972414  -52.8027586  -26.8027586   90.1972414
#>  [56]  -21.8027586  -45.8027586  -36.8027586  -17.8027586   21.1972414
#>  [61]   17.1972414   25.1972414   25.1972414   12.1972414  -12.8027586
#>  [66]   20.1972414   23.1972414  -53.8027586    6.1972414  -31.8027586
#>  [71]  -40.8027586  -46.8027586    9.1972414   13.1972414  -40.8027586
#>  [76]  -22.8027586    1.1972414   14.1972414  -24.8027586   -4.8027586
#>  [81]   20.1972414  -38.8027586   24.1972414    8.1972414  -45.8027586
#>  [86]    3.1972414   28.1972414  -26.8027586   -8.8027586   47.1972414
#>  [91]   -5.8027586   -0.8027586  -15.8027586  -40.8027586  -55.8027586
#>  [96]   41.1972414   -3.8027586   37.1972414  -25.8027586   32.1972414
#> [101]  -10.8027586  -28.8027586  -41.8027586  -44.8027586   25.1972414
#> [106]   65.1972414   41.1972414   -8.8027586  -21.8027586  -12.8027586
#> [111]   -7.8027586   -5.8027586  -27.8027586    0.1972414   16.1972414
#> [116]   20.1972414  -42.8027586    7.1972414    2.1972414    4.1972414
#> [121]  -35.8027586   10.1972414  -18.8027586   -2.8027586  -21.8027586
#> [126]  -23.8027586  -25.8027586   51.1972414  -39.8027586           NA
#> [131]   38.1972414   30.1972414  -17.8027586  -72.8027586   24.1972414
#> [136]   69.1972414    8.1972414   16.1972414  -37.8027586    4.1972414
#> [141]   48.1972414   25.1972414   22.1972414   -6.8027586   35.1972414
#> [146]   10.1972414   30.1972414  -26.8027586  -11.8027586   12.1972414
#> [151]  -21.8027586  -12.8027586    7.1972414  -44.8027586  -47.8027586
#> [156]  -10.8027586   15.1972414           NA  -47.8027586  -24.8027586
#> [161]   29.1972414   10.1972414    5.1972414   12.1972414   14.1972414
#> [166]  -10.8027586   18.1972414    0.1972414   91.1972414   19.1972414
#> [171]    1.1972414  -49.8027586           NA  -74.8027586    6.1972414
#> [176]  -11.8027586  -64.8027586   23.1972414   11.1972414    1.1972414
#> [181]    7.1972414    7.1972414   18.1972414  -16.8027586   26.1972414
#> [186]  -30.8027586    3.1972414           NA    2.1972414  -26.8027586
#> [191]   13.1972414   16.1972414  -22.8027586  -11.8027586   59.1972414
#> [196]   -1.8027586  -27.8027586  -24.8027586   65.1972414   -7.8027586
#> [201]  -28.8027586   -8.8027586  -30.8027586    9.1972414   80.1972414
#> [206]   24.1972414   -3.8027586   14.1972414  -34.8027586   -4.8027586
#> [211]   34.1972414   36.1972414   62.1972414  -30.8027586   38.1972414
#> [216]    4.1972414   28.1972414           NA   10.1972414  -30.8027586
#> [221]  121.1972414   18.1972414  -42.8027586   -3.8027586   31.1972414
#> [226]  -18.8027586    5.1972414   14.1972414   74.1972414  -30.8027586
#> [231]   20.1972414  -23.8027586   11.1972414   24.1972414   22.1972414
#> [236]  -53.8027586   16.1972414  -34.8027586   27.1972414   82.1972414
#> [241]   -7.8027586   24.1972414   25.1972414    1.1972414  -25.8027586
#> [246]  -11.8027586  -13.8027586    3.1972414   53.1972414    8.1972414
#> [251]  -10.8027586  -17.8027586  -26.8027586   81.1972414  -48.8027586
#> [256]   15.1972414    4.1972414   36.1972414    1.1972414  -33.8027586
#> [261]           NA  -25.8027586   61.1972414  -12.8027586    2.1972414
#> [266]   57.1972414           NA  -15.8027586  -22.8027586   30.1972414
#> [271]  -26.8027586   29.1972414   55.1972414   -7.8027586    6.1972414
#> [276]   22.1972414  -35.8027586    5.1972414   12.1972414  -16.8027586
#> [281]   10.1972414   13.1972414  -34.8027586   50.1972414   22.1972414
#> [286]  -38.8027586  -12.8027586   29.1972414   59.1972414   11.1972414
#> [291]   11.1972414   43.1972414   80.1972414  -37.8027586   13.1972414
#> [296]   56.1972414   11.1972414   59.1972414   -9.8027586  -11.8027586
#> [301]  -56.8027586   -6.8027586  -43.8027586    4.1972414  -15.8027586
#> [306]   36.1972414  -12.8027586  -32.8027586   -7.8027586   70.1972414
#> [311]  -47.8027586  -14.8027586  -54.8027586   15.1972414  -13.8027586
#> [316]   69.1972414   -9.8027586   18.1972414   47.1972414   27.1972414
#> [321]   10.1972414   56.1972414  -35.8027586    2.1972414   62.1972414
#> [326]   58.1972414   -4.8027586   -0.8027586  -12.8027586    8.1972414
#> [331]   14.1972414  -45.8027586   45.1972414  -44.8027586  -13.8027586
#> [336]   -9.8027586   32.1972414  -19.8027586   32.1972414  -34.8027586
#> [341]    6.1972414  -37.8027586  -12.8027586   21.1972414   16.1972414
#> [346]  -58.8027586   35.1972414   52.1972414  -17.8027586  -38.8027586
#> [351]   11.1972414   34.1972414    1.1972414  -29.8027586  -15.8027586
#> [356]   49.1972414   45.1972414    5.1972414  -32.8027586  -17.8027586
#> [361]   -7.8027586   -9.8027586  -68.8027586    1.1972414   55.1972414
#> [366]  -17.8027586    7.1972414   16.1972414    8.1972414    0.1972414
#> [371]   38.1972414  -48.8027586  -28.8027586  -70.8027586  -32.8027586
#> [376]  -57.8027586   10.1972414   -1.8027586    7.1972414   -4.8027586
#> [381]  -45.8027586   17.1972414   16.1972414  -48.8027586  -53.8027586
#> [386]  -36.8027586  -48.8027586  -32.8027586   75.1972414   55.1972414
#> [391]  -55.8027586  -52.8027586  -20.8027586    8.1972414  -18.8027586
#> [396]   94.1972414   40.1972414    7.1972414   99.1972414    6.1972414
#> [401]   17.1972414  -41.8027586   33.1972414  -11.8027586  -50.8027586
#> [406]  -14.8027586  -11.8027586   44.1972414  -16.8027586   32.1972414
#> [411]   47.1972414  -31.8027586   27.1972414    0.1972414    6.1972414
#> [416]  -10.8027586  -35.8027586    2.1972414   19.1972414   18.1972414
#> [421]   13.1972414   21.1972414  -44.8027586   69.1972414   35.1972414
#> [426]  -28.8027586  -19.8027586  -73.8027586   20.1972414    2.1972414
#> [431]   -3.8027586    7.1972414  -14.8027586   30.1972414  -16.8027586
#> [436]   13.1972414  -11.8027586 -114.8027586    0.1972414   -4.8027586
#> [441]    7.1972414  -86.8027586  -19.8027586   11.1972414   30.1972414
#> [446]   36.1972414  -15.8027586   -3.8027586   21.1972414   43.1972414
#> [451]   40.1972414  -31.8027586  -33.8027586   -9.8027586  -29.8027586
#> [456]  -36.8027586   13.1972414   61.1972414   47.1972414  -35.8027586
#> [461]  -23.8027586   18.1972414  -35.8027586   42.1972414  -14.8027586
#> [466]  -24.8027586  -23.8027586  -55.8027586   63.1972414  -21.8027586
#> [471]  -65.8027586  -42.8027586    0.1972414  -63.8027586   -4.8027586
#> [476]  -12.8027586  -23.8027586   12.1972414   -1.8027586   78.1972414
#> [481]   -5.8027586  -11.8027586   25.1972414  -43.8027586  -31.8027586
#> [486]    7.1972414  -88.8027586   30.1972414   45.1972414  -23.8027586
#> [491]  -57.8027586   -9.8027586   39.1972414   22.1972414   15.1972414
#> [496]  -17.8027586   48.1972414   -5.8027586  -38.8027586    3.1972414
#> [501]   23.1972414   24.1972414  -13.8027586   32.1972414   -2.8027586
#> [506]           NA   90.1972414   14.1972414   35.1972414   25.1972414
#> [511]    6.1972414    6.1972414    0.1972414   26.1972414  -35.8027586
#> [516]  -30.8027586   10.1972414   17.1972414  -12.8027586  -21.8027586
#> [521]  -48.8027586  -26.8027586  -20.8027586    1.1972414   15.1972414
#> [526]  -57.8027586   -4.8027586   36.1972414  -17.8027586  -12.8027586
#> [531]  -16.8027586  -27.8027586   60.1972414   -4.8027586   41.1972414
#> [536]  -64.8027586    7.1972414   22.1972414   18.1972414   21.1972414
#> [541]  -15.8027586  -21.8027586   31.1972414    4.1972414           NA
#> [546]    7.1972414  -22.8027586   16.1972414    4.1972414   26.1972414
#> [551]  -30.8027586    5.1972414   23.1972414  -38.8027586   29.1972414
#> [556]  -33.8027586   28.1972414           NA   -0.8027586   12.1972414
#> [561]   -4.8027586   17.1972414  -13.8027586   -7.8027586  -37.8027586
#> [566]   -2.8027586   29.1972414  -79.8027586   14.1972414  -13.8027586
#> [571]    6.1972414   -1.8027586   25.1972414  -48.8027586   33.1972414
#> [576]  -14.8027586    5.1972414  -14.8027586   -2.8027586   18.1972414
#> [581]  -50.8027586  -39.8027586  -16.8027586   90.1972414   -4.8027586
#> [586]  -42.8027586   16.1972414    2.1972414   -3.8027586   30.1972414
#> [591]   -4.8027586   42.1972414  -61.8027586  -16.8027586  -15.8027586
#> [596]  -47.8027586   39.1972414   33.1972414   -3.8027586   27.1972414
#> [601]  -36.8027586   38.1972414  -25.8027586  -31.8027586  -47.8027586
#> [606]   -4.8027586   41.1972414  -20.8027586    9.1972414   -3.8027586
#> [611]  -42.8027586  -10.8027586  -23.8027586   43.1972414  -19.8027586
#> [616]  -33.8027586   35.1972414   21.1972414   30.1972414   22.1972414
#> [621]  -10.8027586  -25.8027586   56.1972414  -31.8027586   20.1972414
#> [626]   71.1972414   -4.8027586    5.1972414    7.1972414   -1.8027586
#> [631]  -33.8027586   30.1972414  -26.8027586  -25.8027586  -39.8027586
#> [636]    4.1972414   46.1972414   -9.8027586   24.1972414    8.1972414
#> [641]  -25.8027586   92.1972414   30.1972414   -1.8027586  -13.8027586
#> [646]   35.1972414  -17.8027586  -22.8027586  -50.8027586   -9.8027586
#> [651]   12.1972414   43.1972414   26.1972414   -1.8027586  -26.8027586
#> [656]  -42.8027586  -32.8027586   42.1972414   15.1972414  -11.8027586
#> [661]  -34.8027586   -2.8027586  -48.8027586   26.1972414    6.1972414
#> [666]  -20.8027586  -29.8027586    7.1972414   45.1972414   13.1972414
#> [671]  -39.8027586    2.1972414  -23.8027586   10.1972414   62.1972414
#> [676]    2.1972414   57.1972414  -36.8027586  -46.8027586  -19.8027586
#> [681]   45.1972414  -13.8027586   10.1972414  -29.8027586  -13.8027586
#> [686]  -48.8027586  -16.8027586    5.1972414  -34.8027586   -9.8027586
#> [691]  -41.8027586    6.1972414   52.1972414   42.1972414  -56.8027586
#> [696]   20.1972414  -33.8027586   12.1972414   -5.8027586  -15.8027586
#> [701]   -8.8027586  -36.8027586   13.1972414  -22.8027586   25.1972414
#> [706]  -40.8027586   -2.8027586   34.1972414  -13.8027586   -4.8027586
#> [711]   35.1972414  -32.8027586   27.1972414   30.1972414   -5.8027586
#> [716]  -54.8027586    2.1972414   19.1972414   15.1972414  -25.8027586
#> [721]   55.1972414  -68.8027586   56.1972414    7.1972414  -32.8027586
#> [726]   12.1972414  -15.8027586  -39.8027586   22.1972414    1.1972414
#> [731]   -1.8027586   18.1972414   -3.8027586   29.1972414   21.1972414

# Use a polynomial in regress
regress("mean", atrophy ~ polynomial(age, degree = 2), data = mri)
#> 
#> Call:
#> regress(fnctl = "mean", formula = atrophy ~ polynomial(age, degree = 2), 
#>     data = mri)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -36.828  -9.063  -1.018   7.816  50.887 
#> 
#> Coefficients:
#>                                    Estimate  Naive SE  Robust SE    95%L      
#> [1] Intercept                        35.34     0.5598    0.5622        34.24  
#>     polynomial(age, degree = 2)                                               
#> [2]    age^1                         0.5869    0.1009   0.09808       0.3943  
#> [3]    age^2                        0.02159   0.01099   0.01150     -9.897e-04
#>                                    95%H         F stat    df Pr(>F)   
#> [1] Intercept                        36.45        3952.13 1  < 0.00005
#>     polynomial(age, degree = 2)                     32.91 2  < 0.00005
#> [2]    age^1                         0.7794         35.80 1  < 0.00005
#> [3]    age^2                        0.04417          3.52 1    0.0609 
#> 
#>  Polynomial terms calculated from age, centered at 74.566 
#> 
#> Residual standard error: 12.33 on 732 degrees of freedom
#> Multiple R-squared:  0.09148,	Adjusted R-squared:  0.089 
#> F-statistic: 32.91 on 2 and 732 DF,  p-value: 2.06e-14
#>