Prediction Intervals for uRegress objects

Produces prediction intervals for objects of class uRegress.

Usage

# S3 method for uRegress
predict(object, interval = "prediction", level = 0.95, ...)

Arguments

object

an object of class uRegress.

interval

Type of interval calculation

level

Tolerance/confidence level

...

other arguments to pass to the appropriate predict function for the class of object$fit. See predict.lm, or predict.glm for more details.

Value

Returns a matrix with the fitted value and prediction interval for the entered X.

See also

Examples

# Loading required libraries
library(survival)
library(sandwich)

# Reading in a dataset
data(mri)

# Linear regression of LDL on age (with robust SE by default)
testReg <- regress ("mean", ldl~age, data = mri)

# 95% Prediction Interval for age 50
predict(testReg)
#> Warning: predictions on current data refer to _future_ responses
#>          fit      lwr      upr
#> 1   126.0344 59.97112 192.0977
#> 2   125.2227 59.10603 191.3393
#> 3   124.4110 57.99308 190.8288
#> 4   126.0344 59.97112 192.0977
#> 5   126.2148 60.12959 192.3000
#> 6   126.0344 59.97112 192.0977
#> 7   125.7638 59.71039 191.8172
#> 8   125.7638 59.71039 191.8172
#> 9   126.4853 60.34429 192.6264
#> 10  126.2148 60.12959 192.3000
#> 11  124.7717 58.51818 191.0253
#> 12  126.1246 60.05189 192.1973
#> 13  125.7638 59.71039 191.8172
#> 14  126.2148 60.12959 192.3000
#> 15  125.4031 59.31976 191.4864
#> 16  126.2148 60.12959 192.3000
#> 17  125.2227 59.10603 191.3393
#> 18  126.0344 59.97112 192.0977
#> 19  126.3952 60.27578 192.5145
#> 20  126.4853 60.34429 192.6264
#> 21  124.9521 58.76248 191.1417
#> 22  126.1246 60.05189 192.1973
#> 23  126.2148 60.12959 192.3000
#> 24  126.1246 60.05189 192.1973
#> 25  124.5012 58.12890 190.8734
#> 26  125.4933 59.42202 191.5645
#> 27  126.1246 60.05189 192.1973
#> 28  124.7717 58.51818 191.0253
#> 29  126.1246 60.05189 192.1973
#> 30  126.2148 60.12959 192.3000
#> 31  126.1246 60.05189 192.1973
#> 32  126.2148 60.12959 192.3000
#> 33  125.9442 59.88728 192.0011
#> 34  125.9442 59.88728 192.0011
#> 35  126.4853 60.34429 192.6264
#> 36  125.3129 59.21443 191.4113
#> 37  125.8540 59.80037 191.9076
#> 38  125.7638 59.71039 191.8172
#> 39  126.2148 60.12959 192.3000
#> 40  126.0344 59.97112 192.0977
#> 41  125.2227 59.10603 191.3393
#> 42  126.2148 60.12959 192.3000
#> 43  126.1246 60.05189 192.1973
#> 44  126.2148 60.12959 192.3000
#> 45  126.3050 60.20422 192.4057
#> 46  125.1325 58.99457 191.2704
#> 47  125.8540 59.80037 191.9076
#> 48  125.8540 59.80037 191.9076
#> 49  126.0344 59.97112 192.0977
#> 50  126.3952 60.27578 192.5145
#> 51  125.9442 59.88728 192.0011
#> 52  124.9521 58.76248 191.1417
#> 53  125.4031 59.31976 191.4864
#> 54  125.6736 59.61734 191.7299
#> 55  124.7717 58.51818 191.0253
#> 56  126.1246 60.05189 192.1973
#> 57  126.4853 60.34429 192.6264
#> 58  125.7638 59.71039 191.8172
#> 59  126.0344 59.97112 192.0977
#> 60  125.7638 59.71039 191.8172
#> 61  125.4933 59.42202 191.5645
#> 62  126.2148 60.12959 192.3000
#> 63  125.8540 59.80037 191.9076
#> 64  126.3952 60.27578 192.5145
#> 65  125.9442 59.88728 192.0011
#> 66  126.2148 60.12959 192.3000
#> 67  125.8540 59.80037 191.9076
#> 68  124.7717 58.51818 191.0253
#> 69  126.2148 60.12959 192.3000
#> 70  126.0344 59.97112 192.0977
#> 71  126.0344 59.97112 192.0977
#> 72  125.6736 59.61734 191.7299
#> 73  126.3952 60.27578 192.5145
#> 74  126.2148 60.12959 192.3000
#> 75  125.9442 59.88728 192.0011
#> 76  125.4031 59.31976 191.4864
#> 77  126.0344 59.97112 192.0977
#> 78  126.3952 60.27578 192.5145
#> 79  126.1246 60.05189 192.1973
#> 80  126.4853 60.34429 192.6264
#> 81  126.1246 60.05189 192.1973
#> 82  126.3952 60.27578 192.5145
#> 83  125.2227 59.10603 191.3393
#> 84  125.0423 58.88006 191.2045
#> 85  125.9442 59.88728 192.0011
#> 86  125.4031 59.31976 191.4864
#> 87  126.3050 60.20422 192.4057
#> 88  125.1325 58.99457 191.2704
#> 89  125.5834 59.52122 191.6457
#> 90  124.9521 58.76248 191.1417
#> 91  125.2227 59.10603 191.3393
#> 92  125.0423 58.88006 191.2045
#> 93  125.3129 59.21443 191.4113
#> 94  125.6736 59.61734 191.7299
#> 95  125.7638 59.71039 191.8172
#> 96  125.4031 59.31976 191.4864
#> 97  125.7638 59.71039 191.8172
#> 98  126.0344 59.97112 192.0977
#> 99  125.7638 59.71039 191.8172
#> 100 125.4031 59.31976 191.4864
#> 101 126.0344 59.97112 192.0977
#> 102 125.6736 59.61734 191.7299
#> 103 126.4853 60.34429 192.6264
#> 104 125.7638 59.71039 191.8172
#> 105 125.4031 59.31976 191.4864
#> 106 124.8619 58.64185 191.0820
#> 107 126.3050 60.20422 192.4057
#> 108 125.4031 59.31976 191.4864
#> 109 124.6815 58.39145 190.9716
#> 110 125.6736 59.61734 191.7299
#> 111 125.2227 59.10603 191.3393
#> 112 126.1246 60.05189 192.1973
#> 113 125.8540 59.80037 191.9076
#> 114 126.2148 60.12959 192.3000
#> 115 126.4853 60.34429 192.6264
#> 116 125.8540 59.80037 191.9076
#> 117 126.4853 60.34429 192.6264
#> 118 126.2148 60.12959 192.3000
#> 119 125.8540 59.80037 191.9076
#> 120 126.1246 60.05189 192.1973
#> 121 126.1246 60.05189 192.1973
#> 122 125.3129 59.21443 191.4113
#> 123 126.4853 60.34429 192.6264
#> 124 126.1246 60.05189 192.1973
#> 125 125.6736 59.61734 191.7299
#> 126 125.2227 59.10603 191.3393
#> 127 125.4933 59.42202 191.5645
#> 128 125.1325 58.99457 191.2704
#> 129 126.1246 60.05189 192.1973
#> 131 125.4031 59.31976 191.4864
#> 132 126.3952 60.27578 192.5145
#> 133 125.6736 59.61734 191.7299
#> 134 126.1246 60.05189 192.1973
#> 135 126.3952 60.27578 192.5145
#> 136 125.6736 59.61734 191.7299
#> 137 126.3050 60.20422 192.4057
#> 138 126.4853 60.34429 192.6264
#> 139 124.8619 58.64185 191.0820
#> 140 125.4933 59.42202 191.5645
#> 141 126.3952 60.27578 192.5145
#> 142 126.3952 60.27578 192.5145
#> 143 125.0423 58.88006 191.2045
#> 144 126.0344 59.97112 192.0977
#> 145 125.7638 59.71039 191.8172
#> 146 126.4853 60.34429 192.6264
#> 147 126.0344 59.97112 192.0977
#> 148 124.8619 58.64185 191.0820
#> 149 125.4031 59.31976 191.4864
#> 150 125.6736 59.61734 191.7299
#> 151 126.5755 60.40973 192.7413
#> 152 126.2148 60.12959 192.3000
#> 153 126.1246 60.05189 192.1973
#> 154 126.1246 60.05189 192.1973
#> 155 125.6736 59.61734 191.7299
#> 156 126.3050 60.20422 192.4057
#> 157 125.9442 59.88728 192.0011
#> 159 126.4853 60.34429 192.6264
#> 160 126.2148 60.12959 192.3000
#> 161 125.3129 59.21443 191.4113
#> 162 126.1246 60.05189 192.1973
#> 163 125.4933 59.42202 191.5645
#> 164 125.8540 59.80037 191.9076
#> 165 124.4110 57.99308 190.8288
#> 166 125.9442 59.88728 192.0011
#> 167 126.2148 60.12959 192.3000
#> 168 126.0344 59.97112 192.0977
#> 169 126.3952 60.27578 192.5145
#> 170 126.3952 60.27578 192.5145
#> 171 125.0423 58.88006 191.2045
#> 172 125.5834 59.52122 191.6457
#> 174 126.3050 60.20422 192.4057
#> 175 125.7638 59.71039 191.8172
#> 176 125.2227 59.10603 191.3393
#> 177 125.7638 59.71039 191.8172
#> 178 126.1246 60.05189 192.1973
#> 179 124.7717 58.51818 191.0253
#> 180 126.0344 59.97112 192.0977
#> 181 125.4031 59.31976 191.4864
#> 182 123.5993 56.63561 190.5629
#> 183 125.9442 59.88728 192.0011
#> 184 125.8540 59.80037 191.9076
#> 185 125.5834 59.52122 191.6457
#> 186 124.4110 57.99308 190.8288
#> 187 125.5834 59.52122 191.6457
#> 189 125.4933 59.42202 191.5645
#> 190 126.3050 60.20422 192.4057
#> 191 125.9442 59.88728 192.0011
#> 192 124.5913 58.26170 190.9210
#> 193 126.3050 60.20422 192.4057
#> 194 125.8540 59.80037 191.9076
#> 195 125.4933 59.42202 191.5645
#> 196 125.7638 59.71039 191.8172
#> 197 124.5012 58.12890 190.8734
#> 198 125.2227 59.10603 191.3393
#> 199 125.1325 58.99457 191.2704
#> 200 126.2148 60.12959 192.3000
#> 201 124.7717 58.51818 191.0253
#> 202 125.6736 59.61734 191.7299
#> 203 126.0344 59.97112 192.0977
#> 204 125.4031 59.31976 191.4864
#> 205 126.0344 59.97112 192.0977
#> 206 126.3050 60.20422 192.4057
#> 207 125.4933 59.42202 191.5645
#> 208 126.1246 60.05189 192.1973
#> 209 125.8540 59.80037 191.9076
#> 210 124.5913 58.26170 190.9210
#> 211 124.5012 58.12890 190.8734
#> 212 124.6815 58.39145 190.9716
#> 213 125.9442 59.88728 192.0011
#> 214 124.5012 58.12890 190.8734
#> 215 125.4933 59.42202 191.5645
#> 216 125.8540 59.80037 191.9076
#> 217 125.9442 59.88728 192.0011
#> 219 126.3952 60.27578 192.5145
#> 220 126.2148 60.12959 192.3000
#> 221 126.2148 60.12959 192.3000
#> 222 126.2148 60.12959 192.3000
#> 223 126.3050 60.20422 192.4057
#> 224 125.9442 59.88728 192.0011
#> 225 125.4031 59.31976 191.4864
#> 226 124.6815 58.39145 190.9716
#> 227 123.9600 57.26880 190.6512
#> 228 125.4933 59.42202 191.5645
#> 229 125.4031 59.31976 191.4864
#> 230 124.9521 58.76248 191.1417
#> 231 125.7638 59.71039 191.8172
#> 232 125.0423 58.88006 191.2045
#> 233 126.3050 60.20422 192.4057
#> 234 125.8540 59.80037 191.9076
#> 235 125.5834 59.52122 191.6457
#> 236 125.4933 59.42202 191.5645
#> 237 124.3208 57.85424 190.7873
#> 238 126.4853 60.34429 192.6264
#> 239 126.1246 60.05189 192.1973
#> 240 126.3952 60.27578 192.5145
#> 241 125.9442 59.88728 192.0011
#> 242 125.8540 59.80037 191.9076
#> 243 125.5834 59.52122 191.6457
#> 244 126.1246 60.05189 192.1973
#> 245 125.7638 59.71039 191.8172
#> 246 126.1246 60.05189 192.1973
#> 247 126.3050 60.20422 192.4057
#> 248 126.3050 60.20422 192.4057
#> 249 125.9442 59.88728 192.0011
#> 250 125.4031 59.31976 191.4864
#> 251 126.1246 60.05189 192.1973
#> 252 126.2148 60.12959 192.3000
#> 253 125.2227 59.10603 191.3393
#> 254 126.2148 60.12959 192.3000
#> 255 124.7717 58.51818 191.0253
#> 256 124.6815 58.39145 190.9716
#> 257 125.9442 59.88728 192.0011
#> 258 125.4933 59.42202 191.5645
#> 259 125.1325 58.99457 191.2704
#> 260 125.4031 59.31976 191.4864
#> 262 126.4853 60.34429 192.6264
#> 263 126.2148 60.12959 192.3000
#> 264 125.7638 59.71039 191.8172
#> 265 124.8619 58.64185 191.0820
#> 266 125.0423 58.88006 191.2045
#> 268 126.1246 60.05189 192.1973
#> 269 126.3050 60.20422 192.4057
#> 270 126.3952 60.27578 192.5145
#> 271 126.2148 60.12959 192.3000
#> 272 126.2148 60.12959 192.3000
#> 273 124.6815 58.39145 190.9716
#> 274 125.4031 59.31976 191.4864
#> 275 125.2227 59.10603 191.3393
#> 276 126.3050 60.20422 192.4057
#> 277 125.9442 59.88728 192.0011
#> 278 126.4853 60.34429 192.6264
#> 279 125.5834 59.52122 191.6457
#> 280 125.2227 59.10603 191.3393
#> 281 126.0344 59.97112 192.0977
#> 282 126.2148 60.12959 192.3000
#> 283 125.7638 59.71039 191.8172
#> 284 126.0344 59.97112 192.0977
#> 285 126.0344 59.97112 192.0977
#> 286 126.0344 59.97112 192.0977
#> 287 126.3050 60.20422 192.4057
#> 288 125.7638 59.71039 191.8172
#> 289 125.7638 59.71039 191.8172
#> 290 126.4853 60.34429 192.6264
#> 291 125.9442 59.88728 192.0011
#> 292 125.0423 58.88006 191.2045
#> 293 126.4853 60.34429 192.6264
#> 294 126.0344 59.97112 192.0977
#> 295 126.0344 59.97112 192.0977
#> 296 126.1246 60.05189 192.1973
#> 297 126.2148 60.12959 192.3000
#> 298 126.4853 60.34429 192.6264
#> 299 126.1246 60.05189 192.1973
#> 300 126.1246 60.05189 192.1973
#> 301 125.8540 59.80037 191.9076
#> 302 124.9521 58.76248 191.1417
#> 303 125.2227 59.10603 191.3393
#> 304 126.1246 60.05189 192.1973
#> 305 125.8540 59.80037 191.9076
#> 306 125.7638 59.71039 191.8172
#> 307 126.3050 60.20422 192.4057
#> 308 126.1246 60.05189 192.1973
#> 309 125.9442 59.88728 192.0011
#> 310 126.1246 60.05189 192.1973
#> 311 126.1246 60.05189 192.1973
#> 312 126.4853 60.34429 192.6264
#> 313 126.3050 60.20422 192.4057
#> 314 124.3208 57.85424 190.7873
#> 315 125.8540 59.80037 191.9076
#> 316 125.7638 59.71039 191.8172
#> 317 125.2227 59.10603 191.3393
#> 318 125.5834 59.52122 191.6457
#> 319 125.5834 59.52122 191.6457
#> 320 126.1246 60.05189 192.1973
#> 321 125.0423 58.88006 191.2045
#> 322 125.6736 59.61734 191.7299
#> 323 126.1246 60.05189 192.1973
#> 324 126.3952 60.27578 192.5145
#> 325 126.0344 59.97112 192.0977
#> 326 125.5834 59.52122 191.6457
#> 327 125.8540 59.80037 191.9076
#> 328 126.0344 59.97112 192.0977
#> 329 126.3952 60.27578 192.5145
#> 330 126.3050 60.20422 192.4057
#> 331 125.4933 59.42202 191.5645
#> 332 126.4853 60.34429 192.6264
#> 333 126.2148 60.12959 192.3000
#> 334 125.3129 59.21443 191.4113
#> 335 126.2148 60.12959 192.3000
#> 336 126.4853 60.34429 192.6264
#> 337 125.6736 59.61734 191.7299
#> 338 126.4853 60.34429 192.6264
#> 339 125.9442 59.88728 192.0011
#> 340 125.9442 59.88728 192.0011
#>  [ reached getOption("max.print") -- omitted 392 rows ]