Extract Residuals from uRegress objects

Extracts residuals (unstandardized, standardized, studentized, or jackknife) from uRegress objects.

Usage

# S3 method for uRegress
residuals(object, type = "", ...)

Arguments

object

an object of class uRegress, as returned by regress.

type

denotes the type of residuals to return. Default value is "", which returns unstandardized residuals. "standardized", "studentized", and "jackknife" return the expected type of residuals.

...

other arguments

Value

Returns the type of residuals requested.

Details

Relies on functionality from the stats package to return residuals from the uRegress object. "studentized" residuals are computed as internally studentized residuals, while "jackknife" computes the externally studentized residuals.

See also

Examples

# Reading in a dataset
data(mri)

# Create a uRegress object, regressing ldl on age
ldlReg <- regress("mean", age~ldl, data=mri)

# Get the studentized residuals
residuals(ldlReg, "studentized")
#>           1           2           3           4           5           6 
#> -0.46765314  1.16416172  2.82933726 -0.50111494 -0.82949726 -0.45583332 
#>           7           8           9          10          11          12 
#>  0.06851531  0.07502001 -1.39060232 -0.84590908  2.10385769 -0.66314161 
#>          13          14          15          16          17          18 
#>  0.11462464 -0.83708725  0.81175029 -0.85280153  1.16193199 -0.47857896 
#>          19          20          21          22          23          24 
#> -1.20181825 -1.37592281  1.73385803 -0.67464717 -0.82703930 -0.68084546 
#>          25          26          27          28          29          30 
#>  2.63225947  0.62510968 -0.64959489  2.08235638 -0.65736620 -0.84817990 
#>          31          32          33          34          35          36 
#> -0.65692765 -0.83494611 -0.28357493 -0.26155352 -1.38177495  0.98623822 
#>          37          38          39          40          41          42 
#> -0.08479850  0.06072635 -0.82663310 -0.46423242  1.19111926 -0.82381725 
#>          43          44          45          46          47          48 
#> -0.63204119 -0.85897178 -1.01197224  1.35889264 -0.08997397 -0.12665977 
#>          49          50          51          52          53          54 
#> -0.48395582 -1.18157676 -0.28921422  1.78728755  0.79220932  0.25136693 
#>          55          56          57          58          59          60 
#>  2.14926611 -0.66494679 -1.41155773  0.06332116 -0.47946898  0.08852724 
#>          61          62          63          64          65          66 
#>  0.63778902 -0.82826382 -0.09343014 -1.20098100 -0.29359304 -0.83032452 
#>          67          68          69          70          71          72 
#> -0.09429503  2.07956935 -0.83622774 -0.48576795 -0.48988239  0.24288092 
#>          73          74          75          76          77          78 
#> -1.20223907 -0.83325144 -0.30605787  0.80416365 -0.47110991 -1.20014959 
#>          79          80          81          82          83          84 
#> -0.66630949 -1.39192444 -0.64664166 -1.22418978  1.19205746  1.55200652 
#>          85          86          87          88          89          90 
#> -0.30831742  0.81522452 -1.01076576  1.35367387  0.44274561  1.75494352 
#>          91          92          93          94          95          96 
#>  1.17860199  1.54802760  0.99074153  0.24541566  0.05510741  0.83261819 
#>          97          98          99         100         101         102 
#>  0.07762672 -0.45583332  0.06808218  0.82836268 -0.47636443  0.25051348 
#>         103         104         105         106         107         108 
#> -1.40951192  0.05986166  0.82511196  1.94909690 -1.00566342  0.81003698 
#>         109         110         111         112         113         114 
#>  2.27410052  0.25737318  1.17775146 -0.65780555 -0.11650787 -0.83881853 
#>         115         116         117         118         119         120 
#> -1.38299048 -0.09559306 -1.41002069 -0.83579951 -0.10340080 -0.65344830 
#>         121         122         123         124         125         126 
#> -0.67137150  1.00196571 -1.39829863 -0.65648992  0.25350549  1.17115285 
#>         127         128         129         131         132         133 
#>  0.61921787  1.38924599 -0.67323808  0.83119011 -1.19370713  0.25522152 
#>         134         135         136         137         138         139 
#> -0.68918289 -1.19607969  0.29388654 -1.01901683 -1.38299048  1.90086772 
#>         140         141         142         143         144         145 
#>  0.63202434 -1.18689884 -1.19568066  1.55850408 -0.47460354  0.09466576 
#>         146         147         148         149         150         151 
#> -1.38546691 -0.45874866  1.90482733  0.80876232  0.26824862 -1.58339907 
#>         152         153         154         155         156         157 
#> -0.84455926 -0.65215672 -0.67559097  0.24245935 -1.02731472 -0.28142132 
#>         159         160         161         162         163         164 
#> -1.41259140 -0.85001569  1.01069615 -0.65087226  0.63246306 -0.09905879 
#>         165         166         167         168         169         170 
#>  2.84036253 -0.29271434 -0.83115578 -0.47154459 -1.17248878 -1.19809651 
#>         171         172         174         175         176         177 
#>  1.54889849  0.42553353 -1.05872065  0.08197870  1.17606759  0.05121846 
#>         178         179         180         181         182         183 
#> -0.64538760  2.10432037 -0.47110991  0.81698563  4.48984149 -0.28013311 
#>         184         185         186         187         189         190 
#> -0.11168621  0.45814479  2.82207404  0.44794828  0.63114924 -1.03465782 
#>         191         192         193         194         195         196 
#> -0.28228177  2.47400564 -1.03279114 -0.10950087  0.65735902  0.07849633 
#>         197         198         199         200         201         202 
#>  2.63935171  1.17075250  1.39674607 -0.84233051  2.08781697  0.25909993 
#>         203         204         205         206         207         208 
#> -0.48531397  0.81787223 -0.43846937 -1.01237683  0.62854250 -0.64917066 
#>         209         210         211         212         213         214 
#> -0.11958697  2.46450634  2.66698736  2.30051951 -0.26155352  2.63833026 
#>         215         216         217         219         220         221 
#>  0.64738881 -0.10253145 -0.27585992 -1.20181825 -0.85280153 -0.79334455 
#>         222         223         224         225         226         227 
#> -0.83115578 -1.04233393 -0.28965047  0.82789514  2.27523627  3.75434314 
#>         228         229         230         231         232         233 
#>  0.63644683  0.84897716  1.71961137  0.08808981  1.53855430 -1.01774782 
#>         234         235         236         237         238         239 
#> -0.09386254  0.45634962  0.60786196  3.02500509 -1.40600044 -0.64372627 
#>         240         241         242         243         244         245 
#> -1.17529015 -0.29139906 -0.09386254  0.45769512 -0.65474704  0.06808218 
#>         246         247         248         249         250         251 
#> -0.66045878 -1.02866655 -1.02115665 -0.26530898  0.81742842 -0.66001453 
#>         252         253         254         255         256         257 
#> -0.84681423  1.16995604 -0.80684232  2.08109243  2.28985997 -0.28617040 
#>         258         259         260         262         263         264 
#>  0.64645887  1.36526136  0.79968155 -1.40161300 -0.81414870  0.07371777 
#>         265         266         268         269         270         271 
#>  1.91661181  1.57642324 -0.66224402 -1.03279114 -1.19370713 -0.85094000 
#>         272         273         274         275         276         277 
#> -0.82663310  2.31125605  0.81046383  1.18382641 -1.01318967 -0.30380887 
#>         278         279         280         281         282         283 
#> -1.38757700  0.45190232  1.17399491 -0.46722360 -0.83325144  0.06418639 
#>         284         285         286         287         288         289 
#> -0.45048669 -0.46211244 -0.48896357 -1.02821466  0.09203152  0.10526184 
#>         290         291         292         293         294         295 
#> -1.38504997 -0.28314354  1.56896444 -1.36029274 -0.48850514 -0.46593833 
#>         296         297         298         299         300         301 
#> -0.63204119 -0.83409675 -1.36699731 -0.65957110 -0.66045878 -0.12932566 
#>         302         303         304         305         306         307 
#>  1.72910252  1.16341291 -0.65344830 -0.11124884  0.09510533 -1.02821466 
#>         308         309         310         311         312         313 
#> -0.66998067 -0.29139906 -0.62662049 -0.67701329 -1.39644294 -1.04831470 
#>         314         315         316         317         318         319 
#>  3.02451177 -0.11037455  0.10970862  1.17690666  0.45456379  0.46772731 
#>         320         321         322         323         324         325 
#> -0.64372627  1.55291170  0.28792489 -0.67137150 -1.20522586 -0.44562715 
#>         326         327         328         329         330         331 
#>  0.47285647 -0.10644751 -0.47197983 -1.21187007 -1.01901683  0.63644683 
#>         332         333         334         335         336         337 
#> -1.41155773 -0.82026716  0.97917744 -0.84500815 -1.39416220  0.27710165 
#>         338         339         340         341         342         343 
#> -1.39876689 -0.27415939 -0.30336032 -0.10166259 -1.03989887 -0.10993763 
#>         344         345         346         347         348         349 
#> -0.64622287  1.37200675  0.78991957  0.46221869  0.83793751  2.27562039 
#>         350         351         352         353         354         355 
#> -0.30515702  0.81876288  0.64553227 -0.47110991  0.80129803 -1.21323920 
#>         356         357         358         359         360         361 
#>  3.59512087 -0.63639999 -0.10209696 -1.22125501  1.72460252 -0.47504288 
#>         362         363         364         365         366         367 
#> -1.57780998 -0.68719699  0.63071285 -0.63243379  1.72460252 -1.38672789 
#>         368         369         370         371         372         373 
#>  0.45367434 -0.10079420 -0.47154459 -0.82302153  1.71328155 -0.66813826 
#>         374         375         376         377         378         379 
#> -0.13557779 -0.11870633  2.07839636 -0.09992627  1.36395766 -0.10122834 
#>         380         381         382         383         384         385 
#>  0.62811073  0.05942936 -1.01524314  0.82100731  0.05813259  0.97578881 
#>         386         387         388         389         390         391 
#>  0.06332116 -1.22920550 -0.48622255  1.77087348  0.10348867 -0.86483368 
#>         392         393         394         395         396         397 
#>  0.05640383 -1.39923688  0.45013937  0.25479207  0.30556541  0.09686518 
#>         398         399         400         401         402         403 
#> -0.83579951  0.86219869 -0.65258645 -0.09689194  0.06115873 -1.00877921 
#>         404         405         406         407         408         409 
#> -1.57872196  0.42512408  0.99115809  0.62510968 -1.00451480 -0.11168621 
#>         410         411         412         413         414         415 
#>  0.46085532  0.46772731  0.61673705  0.09115462 -0.28790761 -1.01986902 
#>         416         417         418         419         420         421 
#> -0.84366462  1.35022294 -1.20522586 -0.64706122 -0.46380733 -0.09862521 
#>         422         423         424         425         426         427 
#> -1.38097297 -0.12400150 -0.25864720  0.46221869 -1.40305965 -0.66404253 
#>         428         429         430         431         432         433 
#>  0.96859250 -1.01400739  1.73297430 -0.10601187 -1.01944230 -0.84545809 
#>         434         435         436         437         438         439 
#>  1.56238713  1.90866983 -0.09862521 -0.10950087  0.39984769 -0.83881853 
#>         440         441         442         443         444         445 
#>  0.07719205  2.83705635  0.04170785 -0.48036144  1.00241395 -0.64248828 
#>         446         447         448         449         450         451 
#>  1.19781242 -0.66224402 -0.47328906 -0.27884782 -0.82104927 -1.00604869 
#>         452         453         454         455         456         457 
#> -1.22077128 -0.30291218  0.62596333  0.98503498 -0.67183684 -0.83325144 
#>         458         459         460         461         462         463 
#>  0.10614958  0.46772731  0.61509769  0.62005067 -0.83115578  0.98266032 
#>         464         465         466         467         468         469 
#> -1.00527935  0.44017324  0.25222151  0.62005067 -0.49685724 -0.62931324 
#>         470         471         472         473         474         475 
#> -0.29756588  0.97141602  0.42841310  0.26300330  0.05165056 -0.84100576 
#>         476         477         478         479         480         481 
#> -0.47724845  0.25264922  1.92113194  0.07849633 -0.07059982 -1.20872781 
#>         482         483         484         485         486         487 
#>  1.54337287 -0.46084698  0.24414714  0.61673705  0.08241446 -0.32821832 
#>         488         489         490         491         492         493 
#>  0.45994943  1.38613882  0.06894849 -1.04984038 -0.10862777 -0.27119507 
#>         494         495         496         497         498         499 
#>  0.27265721  0.45323048 -0.47946898 -0.45130362 -0.29052404 -1.22418978 
#>         500         501         502         503         504         505 
#> -0.10296607 -0.46169008 -1.37978255 -1.21232494 -1.19292778 -0.65648992 
#>         507         508         509         510         511         512 
#>  0.11911799 -0.83283030 -0.82421657 -0.09343014 -0.28530388 -0.28530388 
#>         513         514         515         516         517         518 
#>  1.91573604 -1.19528307 -1.22271544  1.16838004 -0.09992627  0.45411878 
#>         519         520         521         522         523         524 
#>  0.07371777  0.62088641 -0.30967829  0.98623822  1.17236237  1.36526136 
#>         525         526         527         528         529         530 
#> -0.46508427 -0.31378467 -0.10644751  1.93295048  0.80623939  0.44102858 
#>         531         532         533         534         535         536 
#>  0.07198252 -0.66767979 -0.26238621  0.99538985  0.09730555 -0.68522607 
#>         537         538         539         540         541         542 
#> -0.46851392 -0.64580485 -0.83115578 -0.46253534  0.62341141 -0.29756588 
#>         543         544         546         547         548         549 
#> -1.19331673 -1.38800409 -0.28487113 -0.11431374  1.18834027  0.81566329 
#>         550         551         552         553         554         555 
#>  0.82557321 -0.66905775 -1.20393699  1.37528514  0.98148884 -1.01036603 
#>         556         557         559         560         561         562 
#> -0.30291218  0.45904592 -0.47197983 -0.28271249 -0.10644751  0.27044854 
#>         563         564         565         566         567         568 
#> -0.29403295 -0.10775525 -0.30470722  0.44533718 -0.27543433 -0.69269468 
#>         569         570         571         572         573         574 
#>  0.45278719  0.44060065 -0.65258645 -0.83969033  1.92736247 -0.12577285 
#>         575         576         577         578         579         580 
#>  0.09378708 -0.11081162 -0.65301698 -0.66179647 -1.20740347  0.27088956 
#>         581         582         583         584         585         586 
#> -1.41415542  0.42965443 -1.03002986  2.33382503 -0.84100576 -0.30696041 
#>         587         588         589         590         591         592 
#>  2.65767213 -0.28703831 -0.28965047 -0.45874866 -1.02464532  0.46542122 
#>         593         594         595         596         597         598 
#> -0.13155333  0.80665745  0.43974635 -0.67701329 -1.19024518 -0.64125711 
#>         599         600         601         602         603         604 
#> -0.84056625 -0.46000604 -0.48804734 -0.82302153 -0.48305351 -0.30201711 
#>         605         606         607         608         609         610 
#> -0.49311876  0.07719205  1.01644635  0.62130539 -0.28400666  1.17945828 
#>         611         612         613         614         615         616 
#>  2.08300352 -0.66001453 -0.11475223 -0.08566044  0.25436292 -0.85420904 
#>         617         618         619         620         621         622 
#> -0.64044011 -1.01359792  1.56238713 -1.38057449 -0.66001453 -0.29934161 
#>         623         624         625         626         627         628 
#> -0.63204119 -1.22077128 -0.64664166 -0.81044845 -0.65736620 -1.20393699 
#>         629         630         631         632         633         634 
#>  0.08241446  0.81304566 -0.67044342 -0.45874866 -0.85094000 -1.03418920 
#>         635         636         637         638         639         640 
#> -0.85705349  0.08110749 -0.63600007 -1.21051443  0.64094889  0.81742842 
#>         641         642         643         644         645         646 
#> -0.29934161  0.48922077 -0.09126942  0.99668300 -1.02866655 -0.08911071 
#>         647         648         649         650         651         652 
#> -0.11212373  0.80416365 -0.86237157  0.99325889 -1.38463470 -0.63720206 
#>         653         654         655         656         657         658 
#>  1.00928742 -0.28877833  0.43508476 -0.49080379  0.61632613 -1.00527935 
#>         659         660         661         662         663         664 
#> -0.46508427 -0.47680614 -0.11958697 -0.10557637 -0.67748917  0.09071638 
#>         665         666         667         668         669         670 
#>  0.63290255  0.80499105 -1.40354538 -0.28487113 -1.18800149 -0.46593833 
#>         671         672         673         674         675         676 
#>  0.98110069  0.26387422 -0.48215369  0.45101972 -0.81377442 -0.65431333 
#>         677         678         679         680         681         682 
#> -0.99966140 -0.12046834 -0.49265445 -0.84772362 -0.82026716  0.99157584 
#>         683         684         685         686         687         688 
#>  0.26737106  0.61756108 -1.21232494  0.05813259  0.07198252 -0.28573697 
#>         689         690         691         692         693         694 
#> -1.40600044  0.99325889 -1.22567809  0.63290255 -0.44967178  1.56844664 
#>         695         696         697         698         699         700 
#>  1.15868104 -0.64664166 -0.11914656 -0.65001989 -0.47416479 -0.66224402 
#>         701         702         703         704         705         706 
#>  0.07545425 -0.85562634  0.63600103  0.62046817 -0.82826382  1.34836273 
#>         707         708         709         710         711         712 
#>  0.62897503  1.38060685 -1.76330014 -0.65736620 -0.82421657  0.61632613 
#>         713         714         715         716         717         718 
#>  0.09115462 -1.00996751  0.81132048 -0.49638758  2.28388683  0.27133094 
#>         719         720         721         722         723         724 
#>  0.08590458  0.98664167  0.10348867  3.18040948 -1.73600048  0.44970004 
#>         725         726         727         728         729         730 
#> -0.30246444  1.73747824  2.46006203  0.61346985  0.64004203 -0.28747279 
#>         731         732         733         734         735 
#>  0.26213367 -0.64748155  1.36309687 -0.45916725 -0.64622287 

# Get the jackknifed residuals
residuals(ldlReg, "jackknife")
#>           1           2           3           4           5           6 
#> -0.46740032  1.16444825  2.84316367 -0.50085525 -0.82931813 -0.45558344 
#>           7           8           9          10          11          12 
#>  0.06846813  0.07496840 -1.39150244 -0.84574250  2.10886741 -0.66288447 
#>          13          14          15          16          17          18 
#>  0.11454639 -0.83691380  0.81155863 -0.85264050  1.16221379 -0.47832365 
#>          19          20          21          22          23          24 
#> -1.20218826 -1.37677465  1.73627205 -0.67439276 -0.82685837 -0.68059266 
#>          25          26          27          28          29          30 
#>  2.64313405  0.62484611 -0.64933501  2.08718418 -0.65710783 -0.84801513 
#>          31          32          33          34          35          36 
#> -0.65666920 -0.83477104 -0.28339452 -0.26138494 -1.38264589  0.98621955 
#>          37          38          39          40          41          42 
#> -0.08474026  0.06068449 -0.82645188 -0.46398042  1.19146484 -0.82363399 
#>          43          44          45          46          47          48 
#> -0.63177850 -0.85881587 -1.01198912  1.35969003 -0.08991223 -0.12657355 
#>          49          50          51          52          53          54 
#> -0.48369937 -1.18190103 -0.28903086  1.79000983  0.79200509  0.25120402 
#>          55          56          57          58          59          60 
#>  2.15467353 -0.66469006 -1.41252893  0.06327753 -0.47921348  0.08846647 
#>          61          62          63          64          65          66 
#>  0.63752716 -0.82808378 -0.09336607 -1.20134908 -0.29340742 -0.83014600 
#>          67          68          69          70          71          72 
#> -0.09423038  2.08437387 -0.83605364 -0.48551113 -0.48962475  0.24272280 
#>          73          74          75          76          77          78 
#> -1.20261005 -0.83307510 -0.30586596  0.80396696 -0.47085628 -1.20051575 
#>          79          80          81          82          83          84 
#> -0.66605306 -1.39282897 -0.64638125 -1.22461274  1.19240516  1.55352283 
#>          85          86          87          88          89          90 
#> -0.30812438  0.81503522 -1.01078092  1.35445490  0.44249931  1.75747668 
#>          91          92          93          94          95          96 
#>  1.17891971  1.54952676  0.99072889  0.24525610  0.05506941  0.83244137 
#>          97          98          99         100         101         102 
#>  0.07757334 -0.45558344  0.06803530  0.82818272 -0.47610961  0.25035104 
#>         103         104         105         106         107         108 
#> -1.41047607  0.05982039  0.82492963  1.95288595 -1.00567133  0.80984416 
#>         109         110         111         112         113         114 
#>  2.28069871  0.25720691  1.17806731 -0.65754727 -0.11642837 -0.83864641 
#>         115         116         117         118         119         120 
#> -1.38386541 -0.09552753 -1.41098659 -0.83562509 -0.10333003 -0.65318915 
#>         121         122         123         124         125         126 
#> -0.67111627  1.00196844 -1.39922456 -0.65623138  0.25334138  1.17145435 
#>         127         128         129         131         132         133 
#>  0.61895365  1.39014159 -0.67298331  0.83101223 -1.19405856  0.25505644 
#>         134         135         136         137         138         139 
#> -0.68893245 -1.19643653  0.29370077 -1.01904392 -1.38386541  1.90431721 
#>         140         141         142         143         144         145 
#>  0.63176165 -1.18723494 -1.19603659  1.56004862 -0.47434911  0.09460085 
#>         146         147         148         149         150         151 
#> -1.38635000 -0.45849803  1.90830399  0.80856865  0.26807638 -1.58505431 
#>         152         153         154         155         156         157 
#> -0.84439161 -0.65189732 -0.67533679  0.24230147 -1.02735411 -0.28124203 
#>         159         160         161         162         163         164 
#> -1.41356617 -0.84985240  1.01071121 -0.65061262  0.63220043 -0.09899093 
#>         165         166         167         168         169         170 
#>  2.85436760 -0.29252917 -0.83097787 -0.47129085 -1.17279317 -1.19845796 
#>         171         172         174         175         176         177 
#>  1.55040140  0.42529241 -1.05880930  0.08192236  1.17637976  0.05118311 
#>         178         179         180         181         182         183 
#> -0.64512698  2.10933405 -0.47085628  0.81679755  4.55062412 -0.27995450 
#>         184         185         186         187         189         190 
#> -0.11160990  0.45789432  2.83578356  0.44770052  0.63088643 -1.03470835 
#>         191         192         193         194         195         196 
#> -0.28210203  2.48282589 -1.03283882 -0.10942602  0.65710065  0.07844237 
#>         197         198         199         200         201         202 
#>  2.65032477  1.17105314  1.39766676 -0.84216111  2.09269058  0.25893270 
#>         203         204         205         206         207         208 
#> -0.48505724  0.81768477 -0.43822431 -1.01239429  0.62827935 -0.64891071 
#>         209         210         211         212         213         214 
#> -0.11950542  2.47321180  2.67834960  2.30738861 -0.26138494  2.64928911 
#>         215         216         217         219         220         221 
#>  0.64712854 -0.10246127 -0.27568359 -1.20218826 -0.85264050 -0.79314102 
#>         222         223         224         225         226         227 
#> -0.83097787 -1.04239635 -0.28946688  0.82771484  2.28184597  3.78886011 
#>         228         229         230         231         232         233 
#>  0.63618478  0.84881303  1.72194672  0.08802934  1.54001304 -1.01777307 
#>         234         235         236         237         238         239 
#> -0.09379818  0.45609961  0.60759671  3.04222567 -1.40695254 -0.64346536 
#>         240         241         242         243         244         245 
#> -1.17560063 -0.29121457 -0.09379818  0.45744476 -0.65448815  0.06803530 
#>         246         247         248         249         250         251 
#> -0.66020106 -1.02870798 -1.02118689 -0.26513834  0.81724065 -0.65975672 
#>         252         253         254         255         256         257 
#> -0.84664838  1.17025496 -0.80664738  2.08590966  2.29661895 -0.28598863 
#>         258         259         260         262         263         264 
#>  0.64619844  1.36607893  0.79948197 -1.40255014 -0.81395867  0.07366705 
#>         265         266         268         269         270         271 
#>  1.92017008  1.57804705 -0.66198669 -1.03283882 -1.19405856 -0.85077746 
#>         272         273         274         275         276         277 
#> -0.82645188  2.31823717  0.81027130  1.18415567 -1.01320830 -0.30361808 
#>         278         279         280         281         282         283 
#> -1.38846706  0.45165349  1.17430256 -0.46697087 -0.83307510  0.06414217 
#>         284         285         286         287         288         289 
#> -0.45023823 -0.46186096 -0.48870612 -1.02825541  0.09196839  0.10518982 
#>         290         291         292         293         294         295 
#> -1.38593168 -0.28296335  1.57055499 -1.36109455 -0.48824777 -0.46568591 
#>         296         297         298         299         300         301 
#> -0.63177850 -0.83392105 -1.36782041 -0.65931319 -0.66020106 -0.12923769 
#>         302         303         304         305         306         307 
#>  1.73149012  1.16369785 -0.65318915 -0.11117282  0.09504013 -1.02825541 
#>         308         309         310         311         312         313 
#> -0.66972511 -0.29121457 -0.62635710 -0.67675948 -1.39736261 -1.04838655 
#>         314         315         316         317         318         319 
#>  3.04172318 -0.11029912  0.10963364  1.17722066  0.45431424  0.46747446 
#>         320         321         322         323         324         325 
#> -0.64346536  1.55443193  0.28774220 -0.67111627 -1.20560377 -0.44538003 
#>         326         327         328         329         330         331 
#>  0.47260243 -0.10637471 -0.47172600 -1.21226355 -1.01904392  0.63618478 
#>         332         333         334         335         336         337 
#> -1.41252893 -0.82008137  0.97914950 -0.84484086 -1.39507421  0.27692465 
#>         338         339         340         341         342         343 
#> -1.39969439 -0.27398397 -0.30316975 -0.10159299 -1.03995749 -0.10986250 
#>         344         345         346         347         348         349 
#> -0.64596239  1.37284593  0.78971395  0.46196719  0.83776472  2.28223399 
#>         350         351         352         353         354         355 
#> -0.30496555  0.81857604  0.64527168 -0.47085628  0.80109948 -1.21363593 
#>         356         357         358         359         360         361 
#>  3.62518314 -0.63613793 -0.10202707 -1.22167087  1.72696526 -0.47478834 
#>         362         363         364         365         366         367 
#> -1.57944001 -0.68694597  0.63044999 -0.63217116  1.72696526 -1.38761515 
#>         368         369         370         371         372         373 
#>  0.45342503 -0.10072518 -0.47129085 -0.82283770  1.71558242 -0.66788226 
#>         374         375         376         377         378         379 
#> -0.13548572 -0.11862536  2.08319110 -0.09985783  1.36477108 -0.10115902 
#>         380         381         382         383         384         385 
#>  0.62784753  0.05938839 -1.01526474  0.82082205  0.05809251  0.97575648 
#>         386         387         388         389         390         391 
#>  0.06327753 -1.22964067 -0.48596564  1.77349882  0.10341784 -0.86468275 
#>         392         393         394         395         396         397 
#>  0.05636494 -1.40016598  0.44989101  0.25462724  0.30537373  0.09679880 
#>         398         399         400         401         402         403 
#> -0.83562509  0.86204551 -0.65232714 -0.09682554  0.06111658 -1.00879153 
#>         404         405         406         407         408         409 
#> -1.58035610  0.42488308  0.99114600  0.62484611 -1.00452110 -0.11160990 
#>         410         411         412         413         414         415 
#>  0.46060416  0.46747446  0.61647257  0.09109208 -0.28772493 -1.01989737 
#>         416         417         418         419         420         421 
#> -0.84349627  1.35099325 -1.20560377 -0.64680089 -0.46355543 -0.09855764 
#>         422         423         424         425         426         427 
#> -1.38184128 -0.12391703 -0.25848023  0.46196719 -1.40400171 -0.66378559 
#>         428         429         430         431         432         433 
#>  0.96855103 -1.01402720  1.73538339 -0.10593936 -1.01947002 -0.84529115 
#>         434         435         436         437         438         439 
#>  1.56394866  1.91217298 -0.09855764 -0.10942602  0.39961526 -0.83864641 
#>         440         441         442         443         444         445 
#>  0.07713896  2.85100769  0.04167905 -0.48010574  1.00241730 -0.64222716 
#>         446         447         448         449         450         451 
#>  1.19817322 -0.66198669 -0.47303492 -0.27866989 -0.82086404 -1.00605714 
#>         452         453         454         455         456         457 
#> -1.22118598 -0.30272183  0.62569986  0.98501471 -0.67158173 -0.83307510 
#>         458         459         460         461         462         463 
#>  0.10607697  0.46747446  0.61483306  0.61978653 -0.83097787  0.98263692 
#>         464         465         466         467         468         469 
#> -1.00528672  0.43992767  0.25205811  0.61978653 -0.49659830 -0.62905019 
#>         470         471         472         473         474         475 
#> -0.29737823  0.97137812  0.42817108  0.26283393  0.05161493 -0.84083533 
#>         476         477         478         479         480         481 
#> -0.47699342  0.25248558  1.92472181  0.07844237 -0.07055122 -1.20911390 
#>         482         483         484         485         486         487 
#>  1.54485210 -0.46059582  0.24398830  0.61647257  0.08235783 -0.32801570 
#>         488         489         490         491         492         493 
#>  0.45969849  1.38702413  0.06890102 -1.04991466 -0.10855351 -0.27102124 
#>         494         495         496         497         498         499 
#>  0.27248259  0.45298129 -0.47921348 -0.45105494 -0.29034000 -1.22461274 
#>         500         501         502         503         504         505 
#> -0.10289559 -0.46143871 -1.38064697 -1.21271951 -1.19327745 -0.65623138 
#>         507         508         509         510         511         512 
#>  0.11903675 -0.83265364 -0.82403360 -0.09336607 -0.28512255 -0.28512255 
#>         513         514         515         516         517         518 
#>  1.91928821 -1.19563809 -1.22313483  1.16867557 -0.09985783  0.45386935 
#>         519         520         521         522         523         524 
#>  0.07366705  0.62062236 -0.30948458  0.98621955  1.17266648  1.36607893 
#>         525         526         527         528         529         530 
#> -0.46483206 -0.31358895 -0.10637471  1.93662376  0.80604406  0.44078277 
#>         531         532         533         534         535         536 
#>  0.07193298 -0.66742368 -0.26221717  0.99538351  0.09723887 -0.68497448 
#>         537         538         539         540         541         542 
#> -0.46826089 -0.64554430 -0.83097787 -0.46228376  0.62314764 -0.29737823 
#>         543         544         546         547         548         549 
#> -1.19366728 -1.38889557 -0.28469003 -0.11423569  1.18867960  0.81547430 
#>         550         551         552         553         554         555 
#>  0.82539121 -0.66880197 -1.20431191  1.37613492  0.98146391 -1.01038061 
#>         556         557         559         560         561         562 
#> -0.30272183  0.45879522 -0.47172600 -0.28253252 -0.10637471  0.27027511 
#>         563         564         565         566         567         568 
#> -0.29384711 -0.10768157 -0.30451598  0.44509015 -0.27525822 -0.69244528 
#>         569         570         571         572         573         574 
#>  0.45253812  0.44035496 -0.65232714 -0.83951888  1.93099617 -0.12568722 
#>         575         576         577         578         579         580 
#>  0.09372277 -0.11073590 -0.65275775 -0.66153904 -1.20778646  0.27071590 
#>         581         582         583         584         585         586 
#> -1.41513561  0.42941202 -1.03007335  2.34104532 -0.84083533 -0.30676805 
#>         587         588         589         590         591         592 
#>  2.66890222 -0.28685608 -0.28946688 -0.45849803 -1.02468073  0.46516893 
#>         593         594         595         596         597         598 
#> -0.13146389  0.80646239  0.43950091 -0.67675948 -1.19058879 -0.64099580 
#>         599         600         601         602         603         604 
#> -0.84039548 -0.45975510 -0.48779007 -0.82283770 -0.48279725 -0.30182722 
#>         605         606         607         608         609         610 
#> -0.49286051  0.07713896  1.01646970  0.62104138 -0.28382602  1.17977788 
#>         611         612         613         614         615         616 
#>  2.08783674 -0.65975672 -0.11467389 -0.08560161  0.25419833 -0.85404917 
#>         617         618         619         620         621         622 
#> -0.64017867 -1.01361714  1.56394866 -1.38144150 -0.65975672 -0.29915307 
#>         623         624         625         626         627         628 
#> -0.63177850 -1.22118598 -0.64638125 -0.81025591 -0.65710783 -1.20431191 
#>         629         630         631         632         633         634 
#>  0.08235783  0.81285488 -0.67018796 -0.45849803 -0.85077746 -1.03423901 
#>         635         636         637         638         639         640 
#> -0.85689598  0.08105175 -0.63573794 -1.21090472  0.64068753  0.81724065 
#>         641         642         643         644         645         646 
#> -0.29915307  0.48896326 -0.09120681  0.99667842 -1.02870798 -0.08904955 
#>         647         648         649         650         651         652 
#> -0.11204714  0.80396696 -0.86221854  0.99324964 -1.38551505 -0.63694012 
#>         653         654         655         656         657         658 
#>  1.00930046 -0.28859520  0.43484069 -0.49054598  0.61606161 -1.00528672 
#>         659         660         661         662         663         664 
#> -0.46483206 -0.47655122 -0.11950542 -0.10550414 -0.67723549  0.09065414 
#>         665         666         667         668         669         670 
#>  0.63263999  0.80479489 -1.40448909 -0.28469003 -1.18834005 -0.46568591 
#>         671         672         673         674         675         676 
#>  0.98107525  0.26370437 -0.48189761  0.45077112 -0.81358413 -0.65405435 
#>         677         678         679         680         681         682 
#> -0.99966093 -0.12038621 -0.49239629 -0.84755849 -0.82008137  0.99156432 
#>         683         684         685         686         687         688 
#>  0.26719930  0.61729668 -1.21271951  0.05809251  0.07193298 -0.28555542 
#>         689         690         691         692         693         694 
#> -1.40695254  0.99324964 -1.22610466  0.63263999 -0.44942354  1.57003491 
#>         695         696         697         698         699         700 
#>  1.15895599 -0.64638125 -0.11906530 -0.64976010 -0.47391046 -0.66198669 
#>         701         702         703         704         705         706 
#>  0.07540234 -0.85546764  0.63573891  0.62020407 -0.82808378  1.34912728 
#>         707         708         709         710         711         712 
#>  0.62871194  1.38147396 -1.76588142 -0.65710783 -0.82403360  0.61606161 
#>         713         714         715         716         717         718 
#>  0.09109208 -1.00998152  0.81112853 -0.49612872  2.29058458  0.27115704 
#>         719         720         721         722         723         724 
#>  0.08584559  0.98662354  0.10341784  3.20067738 -1.73842646  0.44945180 
#>         725         726         727         728         729         730 
#> -0.30227432  1.73991248  2.46871411  0.61320507  0.63978052 -0.28729033 
#>         731         732         733         734         735 
#>  0.26196477 -0.64722130  1.36390755 -0.45891652 -0.64596239