Prediction Intervals for uRegress objects

Produces prediction intervals for objects of class uRegress.

Usage

# S3 method for class 'uRegress'
predict(object, interval = "prediction", level = 0.95, ...)

Arguments

object

an object of class uRegress.

interval

Type of interval calculation

level

Tolerance/confidence level

...

other arguments to pass to the appropriate predict function for the class of object$fit. See predict.lm, or predict.glm for more details.

Value

Returns a matrix with the fitted value and prediction interval for the entered X.

See also

Examples

# Loading required libraries
library(survival)
library(sandwich)

# Reading in a dataset
data(mri)

# Linear regression of LDL on age (with robust SE by default)
testReg <- regress ("mean", ldl~age, data = mri)

# 95% Prediction Interval for age 50
predict(testReg)
#> Warning: predictions on current data refer to _future_ responses
#>          fit      lwr      upr
#> 1   126.0344 59.97112 192.0977
#> 2   125.2227 59.10603 191.3393
#> 3   124.4110 57.99308 190.8288
#> 4   126.0344 59.97112 192.0977
#> 5   126.2148 60.12959 192.3000
#> 6   126.0344 59.97112 192.0977
#> 7   125.7638 59.71039 191.8172
#> 8   125.7638 59.71039 191.8172
#> 9   126.4853 60.34429 192.6264
#> 10  126.2148 60.12959 192.3000
#> 11  124.7717 58.51818 191.0253
#> 12  126.1246 60.05189 192.1973
#> 13  125.7638 59.71039 191.8172
#> 14  126.2148 60.12959 192.3000
#> 15  125.4031 59.31976 191.4864
#> 16  126.2148 60.12959 192.3000
#> 17  125.2227 59.10603 191.3393
#> 18  126.0344 59.97112 192.0977
#> 19  126.3952 60.27578 192.5145
#> 20  126.4853 60.34429 192.6264
#> 21  124.9521 58.76248 191.1417
#> 22  126.1246 60.05189 192.1973
#> 23  126.2148 60.12959 192.3000
#> 24  126.1246 60.05189 192.1973
#> 25  124.5012 58.12890 190.8734
#> 26  125.4933 59.42202 191.5645
#> 27  126.1246 60.05189 192.1973
#> 28  124.7717 58.51818 191.0253
#> 29  126.1246 60.05189 192.1973
#> 30  126.2148 60.12959 192.3000
#> 31  126.1246 60.05189 192.1973
#> 32  126.2148 60.12959 192.3000
#> 33  125.9442 59.88728 192.0011
#> 34  125.9442 59.88728 192.0011
#> 35  126.4853 60.34429 192.6264
#> 36  125.3129 59.21443 191.4113
#> 37  125.8540 59.80037 191.9076
#> 38  125.7638 59.71039 191.8172
#> 39  126.2148 60.12959 192.3000
#> 40  126.0344 59.97112 192.0977
#> 41  125.2227 59.10603 191.3393
#> 42  126.2148 60.12959 192.3000
#> 43  126.1246 60.05189 192.1973
#> 44  126.2148 60.12959 192.3000
#> 45  126.3050 60.20422 192.4057
#> 46  125.1325 58.99457 191.2704
#> 47  125.8540 59.80037 191.9076
#> 48  125.8540 59.80037 191.9076
#> 49  126.0344 59.97112 192.0977
#> 50  126.3952 60.27578 192.5145
#> 51  125.9442 59.88728 192.0011
#> 52  124.9521 58.76248 191.1417
#> 53  125.4031 59.31976 191.4864
#> 54  125.6736 59.61734 191.7299
#> 55  124.7717 58.51818 191.0253
#> 56  126.1246 60.05189 192.1973
#> 57  126.4853 60.34429 192.6264
#> 58  125.7638 59.71039 191.8172
#> 59  126.0344 59.97112 192.0977
#> 60  125.7638 59.71039 191.8172
#> 61  125.4933 59.42202 191.5645
#> 62  126.2148 60.12959 192.3000
#> 63  125.8540 59.80037 191.9076
#> 64  126.3952 60.27578 192.5145
#> 65  125.9442 59.88728 192.0011
#> 66  126.2148 60.12959 192.3000
#> 67  125.8540 59.80037 191.9076
#> 68  124.7717 58.51818 191.0253
#> 69  126.2148 60.12959 192.3000
#> 70  126.0344 59.97112 192.0977
#> 71  126.0344 59.97112 192.0977
#> 72  125.6736 59.61734 191.7299
#> 73  126.3952 60.27578 192.5145
#> 74  126.2148 60.12959 192.3000
#> 75  125.9442 59.88728 192.0011
#> 76  125.4031 59.31976 191.4864
#> 77  126.0344 59.97112 192.0977
#> 78  126.3952 60.27578 192.5145
#> 79  126.1246 60.05189 192.1973
#> 80  126.4853 60.34429 192.6264
#> 81  126.1246 60.05189 192.1973
#> 82  126.3952 60.27578 192.5145
#> 83  125.2227 59.10603 191.3393
#> 84  125.0423 58.88006 191.2045
#> 85  125.9442 59.88728 192.0011
#> 86  125.4031 59.31976 191.4864
#> 87  126.3050 60.20422 192.4057
#> 88  125.1325 58.99457 191.2704
#> 89  125.5834 59.52122 191.6457
#> 90  124.9521 58.76248 191.1417
#> 91  125.2227 59.10603 191.3393
#> 92  125.0423 58.88006 191.2045
#> 93  125.3129 59.21443 191.4113
#> 94  125.6736 59.61734 191.7299
#> 95  125.7638 59.71039 191.8172
#> 96  125.4031 59.31976 191.4864
#> 97  125.7638 59.71039 191.8172
#> 98  126.0344 59.97112 192.0977
#> 99  125.7638 59.71039 191.8172
#> 100 125.4031 59.31976 191.4864
#> 101 126.0344 59.97112 192.0977
#> 102 125.6736 59.61734 191.7299
#> 103 126.4853 60.34429 192.6264
#> 104 125.7638 59.71039 191.8172
#> 105 125.4031 59.31976 191.4864
#> 106 124.8619 58.64185 191.0820
#> 107 126.3050 60.20422 192.4057
#> 108 125.4031 59.31976 191.4864
#> 109 124.6815 58.39145 190.9716
#> 110 125.6736 59.61734 191.7299
#> 111 125.2227 59.10603 191.3393
#> 112 126.1246 60.05189 192.1973
#> 113 125.8540 59.80037 191.9076
#> 114 126.2148 60.12959 192.3000
#> 115 126.4853 60.34429 192.6264
#> 116 125.8540 59.80037 191.9076
#> 117 126.4853 60.34429 192.6264
#> 118 126.2148 60.12959 192.3000
#> 119 125.8540 59.80037 191.9076
#> 120 126.1246 60.05189 192.1973
#> 121 126.1246 60.05189 192.1973
#> 122 125.3129 59.21443 191.4113
#> 123 126.4853 60.34429 192.6264
#> 124 126.1246 60.05189 192.1973
#> 125 125.6736 59.61734 191.7299
#> 126 125.2227 59.10603 191.3393
#> 127 125.4933 59.42202 191.5645
#> 128 125.1325 58.99457 191.2704
#> 129 126.1246 60.05189 192.1973
#> 131 125.4031 59.31976 191.4864
#> 132 126.3952 60.27578 192.5145
#> 133 125.6736 59.61734 191.7299
#> 134 126.1246 60.05189 192.1973
#> 135 126.3952 60.27578 192.5145
#> 136 125.6736 59.61734 191.7299
#> 137 126.3050 60.20422 192.4057
#> 138 126.4853 60.34429 192.6264
#> 139 124.8619 58.64185 191.0820
#> 140 125.4933 59.42202 191.5645
#> 141 126.3952 60.27578 192.5145
#> 142 126.3952 60.27578 192.5145
#> 143 125.0423 58.88006 191.2045
#> 144 126.0344 59.97112 192.0977
#> 145 125.7638 59.71039 191.8172
#> 146 126.4853 60.34429 192.6264
#> 147 126.0344 59.97112 192.0977
#> 148 124.8619 58.64185 191.0820
#> 149 125.4031 59.31976 191.4864
#> 150 125.6736 59.61734 191.7299
#> 151 126.5755 60.40973 192.7413
#> 152 126.2148 60.12959 192.3000
#> 153 126.1246 60.05189 192.1973
#> 154 126.1246 60.05189 192.1973
#> 155 125.6736 59.61734 191.7299
#> 156 126.3050 60.20422 192.4057
#> 157 125.9442 59.88728 192.0011
#> 159 126.4853 60.34429 192.6264
#> 160 126.2148 60.12959 192.3000
#> 161 125.3129 59.21443 191.4113
#> 162 126.1246 60.05189 192.1973
#> 163 125.4933 59.42202 191.5645
#> 164 125.8540 59.80037 191.9076
#> 165 124.4110 57.99308 190.8288
#> 166 125.9442 59.88728 192.0011
#> 167 126.2148 60.12959 192.3000
#> 168 126.0344 59.97112 192.0977
#> 169 126.3952 60.27578 192.5145
#> 170 126.3952 60.27578 192.5145
#> 171 125.0423 58.88006 191.2045
#> 172 125.5834 59.52122 191.6457
#> 174 126.3050 60.20422 192.4057
#> 175 125.7638 59.71039 191.8172
#> 176 125.2227 59.10603 191.3393
#> 177 125.7638 59.71039 191.8172
#> 178 126.1246 60.05189 192.1973
#> 179 124.7717 58.51818 191.0253
#> 180 126.0344 59.97112 192.0977
#> 181 125.4031 59.31976 191.4864
#> 182 123.5993 56.63561 190.5629
#> 183 125.9442 59.88728 192.0011
#> 184 125.8540 59.80037 191.9076
#> 185 125.5834 59.52122 191.6457
#> 186 124.4110 57.99308 190.8288
#> 187 125.5834 59.52122 191.6457
#> 189 125.4933 59.42202 191.5645
#> 190 126.3050 60.20422 192.4057
#> 191 125.9442 59.88728 192.0011
#> 192 124.5913 58.26170 190.9210
#> 193 126.3050 60.20422 192.4057
#> 194 125.8540 59.80037 191.9076
#> 195 125.4933 59.42202 191.5645
#> 196 125.7638 59.71039 191.8172
#> 197 124.5012 58.12890 190.8734
#> 198 125.2227 59.10603 191.3393
#> 199 125.1325 58.99457 191.2704
#> 200 126.2148 60.12959 192.3000
#> 201 124.7717 58.51818 191.0253
#> 202 125.6736 59.61734 191.7299
#> 203 126.0344 59.97112 192.0977
#> 204 125.4031 59.31976 191.4864
#> 205 126.0344 59.97112 192.0977
#> 206 126.3050 60.20422 192.4057
#> 207 125.4933 59.42202 191.5645
#> 208 126.1246 60.05189 192.1973
#> 209 125.8540 59.80037 191.9076
#> 210 124.5913 58.26170 190.9210
#> 211 124.5012 58.12890 190.8734
#> 212 124.6815 58.39145 190.9716
#> 213 125.9442 59.88728 192.0011
#> 214 124.5012 58.12890 190.8734
#> 215 125.4933 59.42202 191.5645
#> 216 125.8540 59.80037 191.9076
#> 217 125.9442 59.88728 192.0011
#> 219 126.3952 60.27578 192.5145
#> 220 126.2148 60.12959 192.3000
#> 221 126.2148 60.12959 192.3000
#> 222 126.2148 60.12959 192.3000
#> 223 126.3050 60.20422 192.4057
#> 224 125.9442 59.88728 192.0011
#> 225 125.4031 59.31976 191.4864
#> 226 124.6815 58.39145 190.9716
#> 227 123.9600 57.26880 190.6512
#> 228 125.4933 59.42202 191.5645
#> 229 125.4031 59.31976 191.4864
#> 230 124.9521 58.76248 191.1417
#> 231 125.7638 59.71039 191.8172
#> 232 125.0423 58.88006 191.2045
#> 233 126.3050 60.20422 192.4057
#> 234 125.8540 59.80037 191.9076
#> 235 125.5834 59.52122 191.6457
#> 236 125.4933 59.42202 191.5645
#> 237 124.3208 57.85424 190.7873
#> 238 126.4853 60.34429 192.6264
#> 239 126.1246 60.05189 192.1973
#> 240 126.3952 60.27578 192.5145
#> 241 125.9442 59.88728 192.0011
#> 242 125.8540 59.80037 191.9076
#> 243 125.5834 59.52122 191.6457
#> 244 126.1246 60.05189 192.1973
#> 245 125.7638 59.71039 191.8172
#> 246 126.1246 60.05189 192.1973
#> 247 126.3050 60.20422 192.4057
#> 248 126.3050 60.20422 192.4057
#> 249 125.9442 59.88728 192.0011
#> 250 125.4031 59.31976 191.4864
#> 251 126.1246 60.05189 192.1973
#> 252 126.2148 60.12959 192.3000
#> 253 125.2227 59.10603 191.3393
#> 254 126.2148 60.12959 192.3000
#> 255 124.7717 58.51818 191.0253
#> 256 124.6815 58.39145 190.9716
#> 257 125.9442 59.88728 192.0011
#> 258 125.4933 59.42202 191.5645
#> 259 125.1325 58.99457 191.2704
#> 260 125.4031 59.31976 191.4864
#> 262 126.4853 60.34429 192.6264
#> 263 126.2148 60.12959 192.3000
#> 264 125.7638 59.71039 191.8172
#> 265 124.8619 58.64185 191.0820
#> 266 125.0423 58.88006 191.2045
#> 268 126.1246 60.05189 192.1973
#> 269 126.3050 60.20422 192.4057
#> 270 126.3952 60.27578 192.5145
#> 271 126.2148 60.12959 192.3000
#> 272 126.2148 60.12959 192.3000
#> 273 124.6815 58.39145 190.9716
#> 274 125.4031 59.31976 191.4864
#> 275 125.2227 59.10603 191.3393
#> 276 126.3050 60.20422 192.4057
#> 277 125.9442 59.88728 192.0011
#> 278 126.4853 60.34429 192.6264
#> 279 125.5834 59.52122 191.6457
#> 280 125.2227 59.10603 191.3393
#> 281 126.0344 59.97112 192.0977
#> 282 126.2148 60.12959 192.3000
#> 283 125.7638 59.71039 191.8172
#> 284 126.0344 59.97112 192.0977
#> 285 126.0344 59.97112 192.0977
#> 286 126.0344 59.97112 192.0977
#> 287 126.3050 60.20422 192.4057
#> 288 125.7638 59.71039 191.8172
#> 289 125.7638 59.71039 191.8172
#> 290 126.4853 60.34429 192.6264
#> 291 125.9442 59.88728 192.0011
#> 292 125.0423 58.88006 191.2045
#> 293 126.4853 60.34429 192.6264
#> 294 126.0344 59.97112 192.0977
#> 295 126.0344 59.97112 192.0977
#> 296 126.1246 60.05189 192.1973
#> 297 126.2148 60.12959 192.3000
#> 298 126.4853 60.34429 192.6264
#> 299 126.1246 60.05189 192.1973
#> 300 126.1246 60.05189 192.1973
#> 301 125.8540 59.80037 191.9076
#> 302 124.9521 58.76248 191.1417
#> 303 125.2227 59.10603 191.3393
#> 304 126.1246 60.05189 192.1973
#> 305 125.8540 59.80037 191.9076
#> 306 125.7638 59.71039 191.8172
#> 307 126.3050 60.20422 192.4057
#> 308 126.1246 60.05189 192.1973
#> 309 125.9442 59.88728 192.0011
#> 310 126.1246 60.05189 192.1973
#> 311 126.1246 60.05189 192.1973
#> 312 126.4853 60.34429 192.6264
#> 313 126.3050 60.20422 192.4057
#> 314 124.3208 57.85424 190.7873
#> 315 125.8540 59.80037 191.9076
#> 316 125.7638 59.71039 191.8172
#> 317 125.2227 59.10603 191.3393
#> 318 125.5834 59.52122 191.6457
#> 319 125.5834 59.52122 191.6457
#> 320 126.1246 60.05189 192.1973
#> 321 125.0423 58.88006 191.2045
#> 322 125.6736 59.61734 191.7299
#> 323 126.1246 60.05189 192.1973
#> 324 126.3952 60.27578 192.5145
#> 325 126.0344 59.97112 192.0977
#> 326 125.5834 59.52122 191.6457
#> 327 125.8540 59.80037 191.9076
#> 328 126.0344 59.97112 192.0977
#> 329 126.3952 60.27578 192.5145
#> 330 126.3050 60.20422 192.4057
#> 331 125.4933 59.42202 191.5645
#> 332 126.4853 60.34429 192.6264
#> 333 126.2148 60.12959 192.3000
#> 334 125.3129 59.21443 191.4113
#> 335 126.2148 60.12959 192.3000
#> 336 126.4853 60.34429 192.6264
#> 337 125.6736 59.61734 191.7299
#> 338 126.4853 60.34429 192.6264
#> 339 125.9442 59.88728 192.0011
#> 340 125.9442 59.88728 192.0011
#> 341 125.8540 59.80037 191.9076
#> 342 126.3050 60.20422 192.4057
#> 343 125.8540 59.80037 191.9076
#> 344 126.1246 60.05189 192.1973
#> 345 125.1325 58.99457 191.2704
#> 346 125.4031 59.31976 191.4864
#> 347 125.5834 59.52122 191.6457
#> 348 125.4031 59.31976 191.4864
#> 349 124.6815 58.39145 190.9716
#> 350 125.9442 59.88728 192.0011
#> 351 125.4031 59.31976 191.4864
#> 352 125.4933 59.42202 191.5645
#> 353 126.0344 59.97112 192.0977
#> 354 125.4031 59.31976 191.4864
#> 355 126.3952 60.27578 192.5145
#> 356 124.0502 57.41965 190.6808
#> 357 126.1246 60.05189 192.1973
#> 358 125.8540 59.80037 191.9076
#> 359 126.3952 60.27578 192.5145
#> 360 124.9521 58.76248 191.1417
#> 361 126.0344 59.97112 192.0977
#> 362 126.5755 60.40973 192.7413
#> 363 126.1246 60.05189 192.1973
#> 364 125.4933 59.42202 191.5645
#> 365 126.1246 60.05189 192.1973
#> 366 124.9521 58.76248 191.1417
#> 367 126.4853 60.34429 192.6264
#> 368 125.5834 59.52122 191.6457
#> 369 125.8540 59.80037 191.9076
#> 370 126.0344 59.97112 192.0977
#> 371 126.2148 60.12959 192.3000
#> 372 124.9521 58.76248 191.1417
#> 373 126.1246 60.05189 192.1973
#> 374 125.8540 59.80037 191.9076
#> 375 125.8540 59.80037 191.9076
#> 376 124.7717 58.51818 191.0253
#> 377 125.8540 59.80037 191.9076
#> 378 125.1325 58.99457 191.2704
#> 379 125.8540 59.80037 191.9076
#> 380 125.4933 59.42202 191.5645
#> 381 125.7638 59.71039 191.8172
#> 382 126.3050 60.20422 192.4057
#> 383 125.4031 59.31976 191.4864
#> 384 125.7638 59.71039 191.8172
#> 385 125.3129 59.21443 191.4113
#> 386 125.7638 59.71039 191.8172
#> 387 126.3952 60.27578 192.5145
#> 388 126.0344 59.97112 192.0977
#> 389 124.9521 58.76248 191.1417
#> 390 125.7638 59.71039 191.8172
#> 391 126.2148 60.12959 192.3000
#> 392 125.7638 59.71039 191.8172
#> 393 126.4853 60.34429 192.6264
#> 394 125.5834 59.52122 191.6457
#> 395 125.6736 59.61734 191.7299
#> 396 125.6736 59.61734 191.7299
#> 397 125.7638 59.71039 191.8172
#> 398 126.2148 60.12959 192.3000
#> 399 125.4031 59.31976 191.4864
#> 400 126.1246 60.05189 192.1973
#> 401 125.8540 59.80037 191.9076
#> 402 125.7638 59.71039 191.8172
#> 403 126.3050 60.20422 192.4057
#> 404 126.5755 60.40973 192.7413
#> 405 125.5834 59.52122 191.6457
#> 406 125.3129 59.21443 191.4113
#> 407 125.4933 59.42202 191.5645
#> 408 126.3050 60.20422 192.4057
#> 409 125.8540 59.80037 191.9076
#> 410 125.5834 59.52122 191.6457
#> 411 125.5834 59.52122 191.6457
#> 412 125.4933 59.42202 191.5645
#> 413 125.7638 59.71039 191.8172
#> 414 125.9442 59.88728 192.0011
#> 415 126.3050 60.20422 192.4057
#> 416 126.2148 60.12959 192.3000
#> 417 125.1325 58.99457 191.2704
#> 418 126.3952 60.27578 192.5145
#> 419 126.1246 60.05189 192.1973
#> 420 126.0344 59.97112 192.0977
#> 421 125.8540 59.80037 191.9076
#> 422 126.4853 60.34429 192.6264
#> 423 125.8540 59.80037 191.9076
#> 424 125.9442 59.88728 192.0011
#> 425 125.5834 59.52122 191.6457
#> 426 126.4853 60.34429 192.6264
#> 427 126.1246 60.05189 192.1973
#> 428 125.3129 59.21443 191.4113
#> 429 126.3050 60.20422 192.4057
#> 430 124.9521 58.76248 191.1417
#> 431 125.8540 59.80037 191.9076
#> 432 126.3050 60.20422 192.4057
#> 433 126.2148 60.12959 192.3000
#> 434 125.0423 58.88006 191.2045
#> 435 124.8619 58.64185 191.0820
#> 436 125.8540 59.80037 191.9076
#> 437 125.8540 59.80037 191.9076
#> 438 125.5834 59.52122 191.6457
#> 439 126.2148 60.12959 192.3000
#> 440 125.7638 59.71039 191.8172
#> 441 124.4110 57.99308 190.8288
#> 442 125.7638 59.71039 191.8172
#> 443 126.0344 59.97112 192.0977
#> 444 125.3129 59.21443 191.4113
#> 445 126.1246 60.05189 192.1973
#> 446 125.2227 59.10603 191.3393
#> 447 126.1246 60.05189 192.1973
#> 448 126.0344 59.97112 192.0977
#> 449 125.9442 59.88728 192.0011
#> 450 126.2148 60.12959 192.3000
#> 451 126.3050 60.20422 192.4057
#> 452 126.3952 60.27578 192.5145
#> 453 125.9442 59.88728 192.0011
#> 454 125.4933 59.42202 191.5645
#> 455 125.3129 59.21443 191.4113
#> 456 126.1246 60.05189 192.1973
#> 457 126.2148 60.12959 192.3000
#> 458 125.7638 59.71039 191.8172
#> 459 125.5834 59.52122 191.6457
#> 460 125.4933 59.42202 191.5645
#> 461 125.4933 59.42202 191.5645
#> 462 126.2148 60.12959 192.3000
#> 463 125.3129 59.21443 191.4113
#> 464 126.3050 60.20422 192.4057
#> 465 125.5834 59.52122 191.6457
#> 466 125.6736 59.61734 191.7299
#> 467 125.4933 59.42202 191.5645
#> 468 126.0344 59.97112 192.0977
#> 469 126.1246 60.05189 192.1973
#> 470 125.9442 59.88728 192.0011
#> 471 125.3129 59.21443 191.4113
#> 472 125.5834 59.52122 191.6457
#> 473 125.6736 59.61734 191.7299
#> 474 125.7638 59.71039 191.8172
#> 475 126.2148 60.12959 192.3000
#> 476 126.0344 59.97112 192.0977
#> 477 125.6736 59.61734 191.7299
#> 478 124.8619 58.64185 191.0820
#> 479 125.7638 59.71039 191.8172
#> 480 125.8540 59.80037 191.9076
#> 481 126.3952 60.27578 192.5145
#> 482 125.0423 58.88006 191.2045
#> 483 126.0344 59.97112 192.0977
#> 484 125.6736 59.61734 191.7299
#> 485 125.4933 59.42202 191.5645
#> 486 125.7638 59.71039 191.8172
#> 487 125.9442 59.88728 192.0011
#> 488 125.5834 59.52122 191.6457
#> 489 125.1325 58.99457 191.2704
#> 490 125.7638 59.71039 191.8172
#> 491 126.3050 60.20422 192.4057
#> 492 125.8540 59.80037 191.9076
#> 493 125.9442 59.88728 192.0011
#> 494 125.6736 59.61734 191.7299
#> 495 125.5834 59.52122 191.6457
#> 496 126.0344 59.97112 192.0977
#> 497 126.0344 59.97112 192.0977
#> 498 125.9442 59.88728 192.0011
#> 499 126.3952 60.27578 192.5145
#> 500 125.8540 59.80037 191.9076
#> 501 126.0344 59.97112 192.0977
#> 502 126.4853 60.34429 192.6264
#> 503 126.3952 60.27578 192.5145
#> 504 126.3952 60.27578 192.5145
#> 505 126.1246 60.05189 192.1973
#> 507 125.7638 59.71039 191.8172
#> 508 126.2148 60.12959 192.3000
#> 509 126.2148 60.12959 192.3000
#> 510 125.8540 59.80037 191.9076
#> 511 125.9442 59.88728 192.0011
#> 512 125.9442 59.88728 192.0011
#> 513 124.8619 58.64185 191.0820
#> 514 126.3952 60.27578 192.5145
#> 515 126.3952 60.27578 192.5145
#> 516 125.2227 59.10603 191.3393
#> 517 125.8540 59.80037 191.9076
#> 518 125.5834 59.52122 191.6457
#> 519 125.7638 59.71039 191.8172
#> 520 125.4933 59.42202 191.5645
#> 521 125.9442 59.88728 192.0011
#> 522 125.3129 59.21443 191.4113
#> 523 125.2227 59.10603 191.3393
#> 524 125.1325 58.99457 191.2704
#> 525 126.0344 59.97112 192.0977
#> 526 125.9442 59.88728 192.0011
#> 527 125.8540 59.80037 191.9076
#> 528 124.8619 58.64185 191.0820
#> 529 125.4031 59.31976 191.4864
#> 530 125.5834 59.52122 191.6457
#> 531 125.7638 59.71039 191.8172
#> 532 126.1246 60.05189 192.1973
#> 533 125.9442 59.88728 192.0011
#> 534 125.3129 59.21443 191.4113
#> 535 125.7638 59.71039 191.8172
#> 536 126.1246 60.05189 192.1973
#> 537 126.0344 59.97112 192.0977
#> 538 126.1246 60.05189 192.1973
#> 539 126.2148 60.12959 192.3000
#> 540 126.0344 59.97112 192.0977
#> 541 125.4933 59.42202 191.5645
#> 542 125.9442 59.88728 192.0011
#> 543 126.3952 60.27578 192.5145
#> 544 126.4853 60.34429 192.6264
#> 546 125.9442 59.88728 192.0011
#> 547 125.8540 59.80037 191.9076
#> 548 125.2227 59.10603 191.3393
#> 549 125.4031 59.31976 191.4864
#> 550 125.4031 59.31976 191.4864
#> 551 126.1246 60.05189 192.1973
#> 552 126.3952 60.27578 192.5145
#> 553 125.1325 58.99457 191.2704
#> 554 125.3129 59.21443 191.4113
#> 555 126.3050 60.20422 192.4057
#> 556 125.9442 59.88728 192.0011
#> 557 125.5834 59.52122 191.6457
#> 559 126.0344 59.97112 192.0977
#> 560 125.9442 59.88728 192.0011
#> 561 125.8540 59.80037 191.9076
#> 562 125.6736 59.61734 191.7299
#> 563 125.9442 59.88728 192.0011
#> 564 125.8540 59.80037 191.9076
#> 565 125.9442 59.88728 192.0011
#> 566 125.5834 59.52122 191.6457
#> 567 125.9442 59.88728 192.0011
#> 568 126.1246 60.05189 192.1973
#> 569 125.5834 59.52122 191.6457
#> 570 125.5834 59.52122 191.6457
#> 571 126.1246 60.05189 192.1973
#> 572 126.2148 60.12959 192.3000
#> 573 124.8619 58.64185 191.0820
#> 574 125.8540 59.80037 191.9076
#> 575 125.7638 59.71039 191.8172
#> 576 125.8540 59.80037 191.9076
#> 577 126.1246 60.05189 192.1973
#> 578 126.1246 60.05189 192.1973
#> 579 126.3952 60.27578 192.5145
#> 580 125.6736 59.61734 191.7299
#> 581 126.4853 60.34429 192.6264
#> 582 125.5834 59.52122 191.6457
#> 583 126.3050 60.20422 192.4057
#> 584 124.6815 58.39145 190.9716
#> 585 126.2148 60.12959 192.3000
#> 586 125.9442 59.88728 192.0011
#> 587 124.5012 58.12890 190.8734
#> 588 125.9442 59.88728 192.0011
#> 589 125.9442 59.88728 192.0011
#> 590 126.0344 59.97112 192.0977
#> 591 126.3050 60.20422 192.4057
#> 592 125.5834 59.52122 191.6457
#> 593 125.8540 59.80037 191.9076
#> 594 125.4031 59.31976 191.4864
#> 595 125.5834 59.52122 191.6457
#> 596 126.1246 60.05189 192.1973
#> 597 126.3952 60.27578 192.5145
#> 598 126.1246 60.05189 192.1973
#> 599 126.2148 60.12959 192.3000
#> 600 126.0344 59.97112 192.0977
#> 601 126.0344 59.97112 192.0977
#> 602 126.2148 60.12959 192.3000
#> 603 126.0344 59.97112 192.0977
#> 604 125.9442 59.88728 192.0011
#> 605 126.0344 59.97112 192.0977
#> 606 125.7638 59.71039 191.8172
#> 607 125.3129 59.21443 191.4113
#> 608 125.4933 59.42202 191.5645
#> 609 125.9442 59.88728 192.0011
#> 610 125.2227 59.10603 191.3393
#> 611 124.7717 58.51818 191.0253
#> 612 126.1246 60.05189 192.1973
#> 613 125.8540 59.80037 191.9076
#> 614 125.8540 59.80037 191.9076
#> 615 125.6736 59.61734 191.7299
#> 616 126.2148 60.12959 192.3000
#> 617 126.1246 60.05189 192.1973
#> 618 126.3050 60.20422 192.4057
#> 619 125.0423 58.88006 191.2045
#> 620 126.4853 60.34429 192.6264
#> 621 126.1246 60.05189 192.1973
#> 622 125.9442 59.88728 192.0011
#> 623 126.1246 60.05189 192.1973
#> 624 126.3952 60.27578 192.5145
#> 625 126.1246 60.05189 192.1973
#> 626 126.2148 60.12959 192.3000
#> 627 126.1246 60.05189 192.1973
#> 628 126.3952 60.27578 192.5145
#> 629 125.7638 59.71039 191.8172
#> 630 125.4031 59.31976 191.4864
#> 631 126.1246 60.05189 192.1973
#> 632 126.0344 59.97112 192.0977
#> 633 126.2148 60.12959 192.3000
#> 634 126.3050 60.20422 192.4057
#> 635 126.2148 60.12959 192.3000
#> 636 125.7638 59.71039 191.8172
#> 637 126.1246 60.05189 192.1973
#> 638 126.3952 60.27578 192.5145
#> 639 125.4933 59.42202 191.5645
#> 640 125.4031 59.31976 191.4864
#> 641 125.9442 59.88728 192.0011
#> 642 125.5834 59.52122 191.6457
#> 643 125.8540 59.80037 191.9076
#> 644 125.3129 59.21443 191.4113
#> 645 126.3050 60.20422 192.4057
#> 646 125.8540 59.80037 191.9076
#> 647 125.8540 59.80037 191.9076
#> 648 125.4031 59.31976 191.4864
#> 649 126.2148 60.12959 192.3000
#> 650 125.3129 59.21443 191.4113
#> 651 126.4853 60.34429 192.6264
#> 652 126.1246 60.05189 192.1973
#> 653 125.3129 59.21443 191.4113
#> 654 125.9442 59.88728 192.0011
#> 655 125.5834 59.52122 191.6457
#> 656 126.0344 59.97112 192.0977
#> 657 125.4933 59.42202 191.5645
#> 658 126.3050 60.20422 192.4057
#> 659 126.0344 59.97112 192.0977
#> 660 126.0344 59.97112 192.0977
#> 661 125.8540 59.80037 191.9076
#> 662 125.8540 59.80037 191.9076
#> 663 126.1246 60.05189 192.1973
#> 664 125.7638 59.71039 191.8172
#> 665 125.4933 59.42202 191.5645
#> 666 125.4031 59.31976 191.4864
#> 667 126.4853 60.34429 192.6264
#> 668 125.9442 59.88728 192.0011
#> 669 126.3952 60.27578 192.5145
#> 670 126.0344 59.97112 192.0977
#> 671 125.3129 59.21443 191.4113
#> 672 125.6736 59.61734 191.7299
#> 673 126.0344 59.97112 192.0977
#> 674 125.5834 59.52122 191.6457
#> 675 126.2148 60.12959 192.3000
#> 676 126.1246 60.05189 192.1973
#> 677 126.3050 60.20422 192.4057
#> 678 125.8540 59.80037 191.9076
#> 679 126.0344 59.97112 192.0977
#> 680 126.2148 60.12959 192.3000
#> 681 126.2148 60.12959 192.3000
#> 682 125.3129 59.21443 191.4113
#> 683 125.6736 59.61734 191.7299
#> 684 125.4933 59.42202 191.5645
#> 685 126.3952 60.27578 192.5145
#> 686 125.7638 59.71039 191.8172
#> 687 125.7638 59.71039 191.8172
#> 688 125.9442 59.88728 192.0011
#> 689 126.4853 60.34429 192.6264
#> 690 125.3129 59.21443 191.4113
#> 691 126.3952 60.27578 192.5145
#> 692 125.4933 59.42202 191.5645
#> 693 126.0344 59.97112 192.0977
#> 694 125.0423 58.88006 191.2045
#> 695 125.2227 59.10603 191.3393
#> 696 126.1246 60.05189 192.1973
#> 697 125.8540 59.80037 191.9076
#> 698 126.1246 60.05189 192.1973
#> 699 126.0344 59.97112 192.0977
#> 700 126.1246 60.05189 192.1973
#> 701 125.7638 59.71039 191.8172
#> 702 126.2148 60.12959 192.3000
#> 703 125.4933 59.42202 191.5645
#> 704 125.4933 59.42202 191.5645
#> 705 126.2148 60.12959 192.3000
#> 706 125.1325 58.99457 191.2704
#> 707 125.4933 59.42202 191.5645
#> 708 125.1325 58.99457 191.2704
#> 709 126.6657 60.47212 192.8593
#> 710 126.1246 60.05189 192.1973
#> 711 126.2148 60.12959 192.3000
#> 712 125.4933 59.42202 191.5645
#> 713 125.7638 59.71039 191.8172
#> 714 126.3050 60.20422 192.4057
#> 715 125.4031 59.31976 191.4864
#> 716 126.0344 59.97112 192.0977
#> 717 124.6815 58.39145 190.9716
#> 718 125.6736 59.61734 191.7299
#> 719 125.7638 59.71039 191.8172
#> 720 125.3129 59.21443 191.4113
#> 721 125.7638 59.71039 191.8172
#> 722 124.2306 57.71238 190.7488
#> 723 126.6657 60.47212 192.8593
#> 724 125.5834 59.52122 191.6457
#> 725 125.9442 59.88728 192.0011
#> 726 124.9521 58.76248 191.1417
#> 727 124.5913 58.26170 190.9210
#> 728 125.4933 59.42202 191.5645
#> 729 125.4933 59.42202 191.5645
#> 730 125.9442 59.88728 192.0011
#> 731 125.6736 59.61734 191.7299
#> 732 126.1246 60.05189 192.1973
#> 733 125.1325 58.99457 191.2704
#> 734 126.0344 59.97112 192.0977
#> 735 126.1246 60.05189 192.1973